refactor: split routes/service.go by concern
service.go had grown to ~3.4k lines covering validation, ownership proof,
placement, config build, stream screening, pushes, drift, health and failover.
Nothing about that mix is wrong as a package - it is one cohesive domain - but
one file that large is hard to review a change against, and the review called
it out.

The declarations move into files named for what they do, in the same package,
with no import-graph change and no interfaces introduced to talk to ourselves:

  service.go    the Service type, its wiring, and the errors handlers match on
  lifecycle.go  create / verify / delete, and the state machine to 'active'
  build.go      DB rows -> Caddy route JSON for one node
  streams.go    the same for L4 streams, plus destination screening
  push.go       getting config onto a node: generations, locking, drift
  health.go     health probing and automatic failover
  util.go       small shared helpers

Pure code motion: every declaration body is byte-identical to before (verified
by parsing both trees and comparing each declaration), and the package exports
exactly the same 99 top-level declarations. Behaviour is unchanged; the whole
suite passes.
