+++ internal/domain/routes/service.go
@@ -1,37 +1,32 @@
 // Package routes owns the route lifecycle: validation, node placement,
 // DNS pre-check, Caddy push.
+//
+// The package is split by concern, all in one package so nothing here needs an
+// interface just to talk to itself:
+//
+//	service.go    the Service type, its wiring and the errors handlers match on
+//	lifecycle.go  create / verify / delete and the state machine that advances a route
+//	build.go      turning DB rows into Caddy route JSON for one node
+//	streams.go    the same for L4 streams, including destination screening
+//	push.go       getting that config onto a node: generations, locking, drift
+//	health.go     node health probing and automatic failover
+//	util.go       small shared helpers
 package routes
 
 import (
-	"bytes"
 	"context"
-	"crypto/rand"
-	"crypto/sha256"
 	"database/sql"
-	"encoding/hex"
-	"encoding/json"
 	"errors"
 	"fmt"
 	"log/slog"
-	"net"
-	"net/url"
 	"runtime/debug"
 	"sort"
 	"strings"
 	"sync"
 	"time"
 
-	"golang.org/x/sync/errgroup"
-
-	"github.com/host-yt/caddy-proxy-manager/internal/audit"
 	"github.com/host-yt/caddy-proxy-manager/internal/caddyapi"
-	"github.com/host-yt/caddy-proxy-manager/internal/cloudflare"
-	"github.com/host-yt/caddy-proxy-manager/internal/dns"
-	"github.com/host-yt/caddy-proxy-manager/internal/geoip"
 	"github.com/host-yt/caddy-proxy-manager/internal/quota"
-	"github.com/host-yt/caddy-proxy-manager/internal/security"
-	"github.com/host-yt/caddy-proxy-manager/internal/store"
-	"github.com/host-yt/caddy-proxy-manager/internal/streamguard"
 )
 
 // recoverBg logs and swallows a panic in a fire-and-forget goroutine.
@@ -247,172 +242,6 @@ func (s *Service) nodeLock(id int64) *sync.Mutex {
 	return m
 }
 
-// bumpDesiredGen records that nodeID's config changed and returns the new
-// generation. Called before every scheduled push so an in-flight push can tell
-// that the snapshot it built is already out of date.
-func (s *Service) bumpDesiredGen(nodeID int64) uint64 {
-	s.genMu.Lock()
-	defer s.genMu.Unlock()
-	if s.desiredGen == nil {
-		s.desiredGen = map[int64]uint64{}
-	}
-	s.desiredGen[nodeID]++
-	return s.desiredGen[nodeID]
-}
-
-// currentGen returns nodeID's desired generation without bumping it.
-func (s *Service) currentGen(nodeID int64) uint64 {
-	s.genMu.Lock()
-	defer s.genMu.Unlock()
-	return s.desiredGen[nodeID]
-}
-
-// recordApplied marks gen as the generation now live on nodeID.
-func (s *Service) recordApplied(nodeID int64, gen uint64) {
-	s.genMu.Lock()
-	defer s.genMu.Unlock()
-	if s.appliedGen == nil {
-		s.appliedGen = map[int64]uint64{}
-	}
-	if gen > s.appliedGen[nodeID] {
-		s.appliedGen[nodeID] = gen
-	}
-}
-
-// AppliedGeneration reports the config generation last successfully loaded on
-// nodeID. Exported for tests and diagnostics.
-func (s *Service) AppliedGeneration(nodeID int64) uint64 {
-	s.genMu.Lock()
-	defer s.genMu.Unlock()
-	return s.appliedGen[nodeID]
-}
-
-// schedulePush debounces a full-config push to nodeID. Within the debounce
-// window (PushDebounceMs) repeated calls reset the timer; only the last fires.
-// Falls back to an immediate goroutine push when debouncing is disabled (0).
-func (s *Service) schedulePush(nodeID int64) {
-	s.bumpDesiredGen(nodeID)
-	window := time.Duration(s.PushDebounceMs) * time.Millisecond
-	if window <= 0 {
-		go func() {
-			defer recoverBg(s.Logger, "schedulePush.immediate")
-			ctx, cancel := context.WithTimeout(s.BackgroundCtx(), 30*time.Second)
-			defer cancel()
-			if err := s.pushNodeConfig(ctx, nodeID); err != nil && s.Logger != nil {
-				s.Logger.Warn("immediate push failed", "node_id", nodeID, "err", err)
-			}
-		}()
-		return
-	}
-	s.debounceMu.Lock()
-	defer s.debounceMu.Unlock()
-	if s.debouncers == nil {
-		s.debouncers = make(map[int64]*time.Timer)
-	}
-	if t, ok := s.debouncers[nodeID]; ok {
-		t.Reset(window) // coalesce: push further into the future
-		return
-	}
-	s.debouncers[nodeID] = time.AfterFunc(window, func() {
-		s.debounceMu.Lock()
-		delete(s.debouncers, nodeID)
-		s.debounceMu.Unlock()
-		defer recoverBg(s.Logger, "schedulePush.debounced")
-		ctx, cancel := context.WithTimeout(s.BackgroundCtx(), 30*time.Second)
-		defer cancel()
-		if err := s.pushNodeConfig(ctx, nodeID); err != nil && s.Logger != nil {
-			s.Logger.Warn("debounced push failed", "node_id", nodeID, "err", err)
-		}
-	})
-}
-
-// SchedulePush is the exported debounce entry-point for external callers
-// (handlers, wg_bootstrap) that need to trigger a node push after a config
-// change but don't want to import the full push path directly.
-func (s *Service) SchedulePush(nodeID int64) { s.schedulePush(nodeID) }
-
-// SchedulePushAllNodes re-pushes every node that hosts a route. Used when a
-// panel-wide setting baked into generated config (geo-block default, error-page
-// branding) changes and must reach all nodes, not just one client's.
-func (s *Service) SchedulePushAllNodes(ctx context.Context) {
-	if s.DB == nil {
-		return
-	}
-	rows, err := s.DB.QueryContext(ctx,
-		`SELECT DISTINCT caddy_node_id FROM routes WHERE caddy_node_id IS NOT NULL`)
-	if err != nil {
-		return
-	}
-	defer rows.Close()
-	for rows.Next() {
-		var nid int64
-		if rows.Scan(&nid) == nil && nid != 0 {
-			s.schedulePush(nid)
-		}
-	}
-}
-
-// SchedulePushForClient re-pushes every node hosting a route owned by the given
-// client. Used when a client-level setting (e.g. geo-block page) changes and
-// must propagate to all of that client's routes.
-func (s *Service) SchedulePushForClient(ctx context.Context, clientID int64) {
-	if s.DB == nil || clientID == 0 {
-		return
-	}
-	rows, err := s.DB.QueryContext(ctx,
-		`SELECT DISTINCT r.caddy_node_id FROM routes r
-		   JOIN services sv ON sv.id = r.service_id
-		  WHERE sv.client_id = ? AND r.caddy_node_id IS NOT NULL`, clientID)
-	if err != nil {
-		return
-	}
-	defer rows.Close()
-	for rows.Next() {
-		var nid int64
-		if rows.Scan(&nid) == nil && nid != 0 {
-			s.schedulePush(nid)
-		}
-	}
-}
-
-// SchedulePushForRoute re-pushes every node serving the given route (its direct
-// caddy_node_id plus any route_node_assignments fan-out). Used when something
-// off the routes table but baked into a node's config changes - e.g. a manual
-// TLS cert linked to the route is imported, replaced, or deleted.
-func (s *Service) SchedulePushForRoute(ctx context.Context, routeID int64) {
-	if s.DB == nil || routeID == 0 {
-		return
-	}
-	seen := map[int64]struct{}{}
-	sched := func(nid int64) {
-		if nid == 0 {
-			return
-		}
-		if _, dup := seen[nid]; dup {
-			return
-		}
-		seen[nid] = struct{}{}
-		s.schedulePush(nid)
-	}
-	var direct sql.NullInt64
-	if err := s.DB.QueryRowContext(ctx,
-		`SELECT caddy_node_id FROM routes WHERE id = ?`, routeID).Scan(&direct); err == nil && direct.Valid {
-		sched(direct.Int64)
-	}
-	rows, err := s.DB.QueryContext(ctx,
-		`SELECT node_id FROM route_node_assignments WHERE route_id = ?`, routeID)
-	if err != nil {
-		return
-	}
-	defer rows.Close()
-	for rows.Next() {
-		var nid int64
-		if rows.Scan(&nid) == nil {
-			sched(nid)
-		}
-	}
-}
-
 // BackgroundCtx returns the app background context (cancelled after shutdown),
 // or context.Background() when unset (tests/dev). Exported so handlers in other
 // packages can scope their fire-and-forget pushes to the app lifecycle.
@@ -573,2961 +402,4 @@ func (s *Service) ExternalAllowlistAll() []string {
 	}
 	sort.Strings(out)
 	return out
-}
-
-// Create inserts a route, picks a node, runs DNS pre-check synchronously
-// (best-effort), and pushes the node config to Caddy. Returns the new
-// route id.
-func (s *Service) Create(ctx context.Context, clientID int64, in CreateInput) (int64, error) {
-	domain := strings.ToLower(strings.TrimSpace(in.Domain))
-	pathPrefix := strings.TrimSpace(in.PathPrefix)
-	if domain == "" || !validDomain(domain) {
-		return 0, ErrInvalidDomain
-	}
-	if pathPrefix != "" {
-		if !strings.HasPrefix(pathPrefix, "/") {
-			pathPrefix = "/" + pathPrefix
-		}
-		if strings.Contains(pathPrefix, "..") {
-			return 0, ErrInvalidDomain
-		}
-	}
-
-	// Verify service ownership + read port range + node_group + plan.
-	var (
-		backendIP    string
-		portStart    int
-		portEnd      int
-		ownerClient  int64
-		nodeGroupID  int64
-		planSSL      bool
-		planWS       bool
-		planPath     bool
-		planMaxDom   int
-		planWild     bool
-		planExtProxy bool
-	)
-	err := s.DB.QueryRowContext(ctx,
-		`SELECT s.client_id, s.backend_ip, s.allowed_port_start, s.allowed_port_end, s.node_group_id,
-		        p.ssl_enabled, p.websocket_enabled, p.path_routing_enabled, p.max_domains, p.wildcard_enabled,
-		        p.external_proxy_enabled
-		 FROM services s JOIN plans p ON p.id = s.plan_id
-		 WHERE s.id = ? LIMIT 1`,
-		in.ServiceID,
-	).Scan(&ownerClient, &backendIP, &portStart, &portEnd, &nodeGroupID, &planSSL, &planWS, &planPath, &planMaxDom, &planWild, &planExtProxy)
-	if errors.Is(err, sql.ErrNoRows) {
-		return 0, ErrServiceNotYours
-	}
-	if err != nil {
-		return 0, fmt.Errorf("service lookup: %w", err)
-	}
-	// 0 clientID means "called from admin/API context" - allow.
-	if clientID != 0 && ownerClient != clientID {
-		return 0, ErrServiceNotYours
-	}
-	// External-HTTPS-upstream setup (admin-only). Validate the FQDN against
-	// the allowlist (primary open-relay defense) and force the route shape:
-	// https proxy to the origin's :443, its own cert via On-Demand TLS.
-	var externalHost, encSecret string
-	if in.External {
-		// Per-plan gate. Admin/API context (clientID==0) still requires the
-		// plan flag; the admin-self plan is kind='npm' which the migration
-		// sets external_proxy_enabled=1, so super_admin works out of the box.
-		if !planExtProxy {
-			return 0, ErrExternalNotInPlan
-		}
-		externalHost = strings.ToLower(strings.TrimSpace(in.ExternalHost))
-		if !s.externalHostAllowed(externalHost) {
-			return 0, ErrExternalHostNotAllowed
-		}
-		in.Kind = "proxy"
-		in.UpstreamScheme = "https"
-		in.SSL = true
-		if in.UpstreamPort == 0 {
-			in.UpstreamPort = 443
-		}
-		if in.ProxySecretPlain != "" {
-			if s.EncryptSecret == nil {
-				return 0, fmt.Errorf("external route secret encryption not configured")
-			}
-			encSecret, err = s.EncryptSecret(in.ProxySecretPlain)
-			if err != nil {
-				return 0, fmt.Errorf("encrypt proxy secret: %w", err)
-			}
-		}
-	}
-
-	// Redirect routes have no upstream; skip the port-range check and
-	// store port=0 so the column (NOT NULL) stays valid. External routes
-	// target the origin's port (443), not the customer range, so skip too.
-	if in.Kind == "redirect" {
-		in.UpstreamPort = 0
-		if in.RedirectURL == "" {
-			return 0, fmt.Errorf("redirect_url is required for redirect routes")
-		}
-		switch in.RedirectCode {
-		case 0:
-			in.RedirectCode = 308
-		case 301, 302, 307, 308:
-		default:
-			return 0, fmt.Errorf("redirect_code must be 301/302/307/308")
-		}
-	} else if !in.External && (in.UpstreamPort < portStart || in.UpstreamPort > portEnd) {
-		return 0, ErrPortOutOfRange
-	}
-	// Reject a backend port already claimed by another route in this service's pool.
-	if in.Kind != "redirect" && !in.External {
-		var portUsed int
-		if err := s.DB.QueryRowContext(ctx,
-			"SELECT COUNT(*) FROM routes WHERE service_id = ? AND upstream_port = ?",
-			in.ServiceID, in.UpstreamPort,
-		).Scan(&portUsed); err == nil && portUsed > 0 {
-			return 0, ErrPortInUse
-		}
-	}
-	if pathPrefix != "" && !planPath {
-		return 0, fmt.Errorf("plan does not permit path routing")
-	}
-	// Plan limit: max_domains counted across this service.
-	if planMaxDom > 0 {
-		var currentCount int
-		if err := s.DB.QueryRowContext(ctx,
-			"SELECT COUNT(*) FROM routes WHERE service_id = ?", in.ServiceID,
-		).Scan(&currentCount); err == nil && currentCount >= planMaxDom {
-			return 0, ErrMaxDomains
-		}
-	}
-	// Reseller aggregate quota. Single choke point: panel, client portal and
-	// API route creation all pass through here. Lookup errors log + allow
-	// (business limit, must not brick creates on a transient).
-	if s.Quota != nil {
-		rid, qerr := s.Quota.ResellerOfClient(ctx, ownerClient)
-		if qerr == nil && rid != 0 {
-			qerr = s.Quota.CanCreateRoute(ctx, rid, in.ServiceID)
-		}
-		if errors.Is(qerr, quota.ErrDomainQuota) {
-			return 0, qerr
-		}
-		if qerr != nil {
-			s.Logger.Warn("reseller quota check skipped", "client", ownerClient, "err", qerr)
-		}
-	}
-
-	// Plan flags constrain customer choice. External routes always need their
-	// own cert (the node domain), so the plan SSL flag can't disable them.
-	if !planSSL && !in.External {
-		in.SSL = false
-	}
-	if !planWS {
-		in.WebSocket = false
-	}
-
-	// Wildcard DNS-01: plan-gated for customers (admin clientID==0 bypasses,
-	// like External). Require an enabled dns_providers row for the zone (else
-	// the cert can never issue) and that the domain is the zone or a subdomain
-	// of it (else *.zone would not cover it). DNS A/AAAA is still required for
-	// the host data-plane and is checked later in advanceRoute.
-	var wildcardZone string
-	if in.WildcardEnabled {
-		if clientID != 0 && !planWild {
-			return 0, fmt.Errorf("plan does not permit wildcard certificates")
-		}
-		wildcardZone = strings.ToLower(strings.TrimSpace(in.WildcardZone))
-		if wildcardZone == "" || !validDomain(wildcardZone) {
-			return 0, ErrWildcardZoneMismatch
-		}
-		if domain != wildcardZone && !strings.HasSuffix(domain, "."+wildcardZone) {
-			return 0, ErrWildcardZoneMismatch
-		}
-		var n int
-		if err := s.DB.QueryRowContext(ctx,
-			"SELECT COUNT(*) FROM dns_providers WHERE name = ?", wildcardZone).Scan(&n); err != nil || n == 0 {
-			return 0, ErrWildcardNoProvider
-		}
-	}
-
-	// Pick node(s) based on group mode: single / active_active / failover.
-	// Primary slot lands in routes.caddy_node_id; for fan-out modes the
-	// other nodes get rows in route_node_assignments after insert.
-	var (
-		primaryNode int64
-		allNodes    []int64
-		groupMode   string
-	)
-	primaryNode, allNodes, groupMode, err = nodePlacement(ctx, s.DB, nodeGroupID)
-	if err != nil {
-		if errors.Is(err, sql.ErrNoRows) {
-			return 0, ErrNoNodeFound
-		}
-		return 0, fmt.Errorf("node placement: %w", err)
-	}
-	nodeID := primaryNode
-
-	// Tunnel guard: peer must belong to the owning client and cover the
-	// placed node (directly or via its peer group), else Caddy would dial
-	// through a nonexistent wg interface. Returns peer IP for the override
-	// dedupe below.
-	var tunnelPeerIP string
-	if in.ViaWGPeerID > 0 {
-		if err := s.DB.QueryRowContext(ctx,
-			`SELECT COALESCE(p_base.assigned_ip,'') FROM customer_wg_peer p_base
-			  WHERE p_base.id = ? AND p_base.client_id = ? AND p_base.status <> 'revoked'
-			    AND EXISTS (SELECT 1 FROM customer_wg_peer p_use
-			                 WHERE p_use.status <> 'revoked' AND p_use.node_id = ?
-			                   AND (p_use.id = p_base.id OR
-			                        (p_base.peer_group_id IS NOT NULL AND p_use.peer_group_id = p_base.peer_group_id)))`,
-			in.ViaWGPeerID, ownerClient, nodeID).Scan(&tunnelPeerIP); err != nil {
-			return 0, ErrTunnelNotOnNode
-		}
-	}
-
-	// Insert + increment counter in a transaction.
-	tx, err := s.DB.BeginTx(ctx, nil)
-	if err != nil {
-		return 0, err
-	}
-	defer tx.Rollback() //nolint:errcheck
-
-	kind := in.Kind
-	if kind != "redirect" {
-		kind = "proxy"
-	}
-	var redirURL sql.NullString
-	if in.RedirectURL != "" {
-		redirURL = sql.NullString{String: in.RedirectURL, Valid: true}
-	}
-	var redirCode sql.NullInt32
-	if in.RedirectCode != 0 {
-		redirCode = sql.NullInt32{Int32: int32(in.RedirectCode), Valid: true}
-	}
-	var tagVal sql.NullString
-	if t := strings.TrimSpace(in.Tag); t != "" {
-		if len(t) > 64 {
-			t = t[:64]
-		}
-		tagVal = sql.NullString{String: t, Valid: true}
-	}
-	scheme := in.UpstreamScheme
-	if scheme != "https" {
-		scheme = "http"
-	}
-	// External-route columns (NULL/0 for normal routes).
-	var backendOverride, hostHeader, secretEnc sql.NullString
-	extFlag := 0
-	// Tunnel route: per-route override beats peer IP in the build COALESCE;
-	// hostname backends also get the peer as DNS resolver (dynamic upstreams).
-	// No override when backend equals peer IP - fan-out nodes must pick their
-	// own group member's IP.
-	var viaPeer, dnsResolverPeer sql.NullInt64
-	if in.ViaWGPeerID > 0 && kind == "proxy" && !in.External {
-		viaPeer = sql.NullInt64{Int64: in.ViaWGPeerID, Valid: true}
-		if backendIP != "" && backendIP != tunnelPeerIP {
-			backendOverride = sql.NullString{String: backendIP, Valid: true}
-			if !looksLikeIP(backendIP) {
-				dnsResolverPeer = viaPeer
-			}
-		}
-	}
-	if in.External {
-		extFlag = 1
-		backendOverride = sql.NullString{String: externalHost, Valid: true}
-		hh := strings.TrimSpace(in.UpstreamHostHeader)
-		if hh == "" {
-			hh = externalHost
-		}
-		hostHeader = sql.NullString{String: hh, Valid: true}
-		if encSecret != "" {
-			secretEnc = sql.NullString{String: encSecret, Valid: true}
-		}
-	}
-	// Wildcard columns (0/NULL for normal routes).
-	var wildFlag int
-	var wildZone sql.NullString
-	if in.WildcardEnabled {
-		wildFlag = 1
-		wildZone = sql.NullString{String: wildcardZone, Valid: true}
-	}
-	// Domain-ownership gate: admin/API context (clientID==0) is trusted and lands
-	// verified; self-service routes land unverified with a token the owner must
-	// publish as a DNS TXT record before the route can advance or get a cert.
-	verified := 0
-	verifyToken := ""
-	if clientID == 0 {
-		verified = 1
-	} else {
-		verifyToken, err = newVerifyToken()
-		if err != nil {
-			return 0, fmt.Errorf("verify token: %w", err)
-		}
-	}
-	// Anti-squat: the UNIQUE(domain,path) constraint would otherwise let a
-	// squatter's first-come UNVERIFIED row permanently block a later claim for the
-	// same domain. If the sole conflicting row is unverified, evict it inside this
-	// tx so the new claim can proceed (both then race to prove ownership via TXT).
-	// A VERIFIED conflicting row is never displaced - that owner proved control.
-	var conflictID, conflictVerified int64
-	var conflictNode sql.NullInt64
-	if e := tx.QueryRowContext(ctx,
-		`SELECT id, domain_verified, caddy_node_id FROM routes
-		 WHERE domain = ? AND COALESCE(path_prefix,'') = ? LIMIT 1`,
-		domain, pathPrefix,
-	).Scan(&conflictID, &conflictVerified, &conflictNode); e == nil {
-		if conflictVerified == 1 {
-			return 0, ErrDomainTaken
-		}
-		if _, e := tx.ExecContext(ctx, "DELETE FROM routes WHERE id = ? AND domain_verified = 0", conflictID); e != nil {
-			return 0, ErrDomainTaken
-		}
-		if conflictNode.Valid && conflictNode.Int64 != 0 {
-			_, _ = tx.ExecContext(ctx,
-				"UPDATE caddy_nodes SET current_routes = GREATEST(current_routes - 1, 0) WHERE id = ?", conflictNode.Int64)
-		}
-		s.Logger.Warn("anti-squat: evicted unverified route on create", "evicted_id", conflictID, "domain", domain)
-	}
-	res, err := tx.ExecContext(ctx,
-		`INSERT INTO routes (service_id, caddy_node_id, domain, path_prefix, upstream_port, upstream_scheme,
-		   ssl_enabled, websocket, force_https, http2_enabled, http3_enabled, status,
-		   kind, redirect_url, redirect_code, tag,
-		   backend_ip_override, upstream_external, upstream_host_header, proxy_secret_enc,
-		   wildcard_enabled, wildcard_zone, group_id, custom_fields,
-		   via_wg_peer_id, dns_resolver_via_wg_peer_id,
-		   domain_verified, verify_token)
-		 VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, 1, 0, 'pending_dns', ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, NULLIF(?, 0), NULLIF(?, ''), ?, ?, ?, ?)`,
-		in.ServiceID, nodeID, domain, pathPrefix, in.UpstreamPort, scheme,
-		in.SSL, in.WebSocket, in.ForceHTTPS,
-		kind, redirURL, redirCode, tagVal,
-		backendOverride, extFlag, hostHeader, secretEnc,
-		wildFlag, wildZone, in.GroupID, in.CustomFields,
-		viaPeer, dnsResolverPeer,
-		verified, verifyToken)
-	if err != nil {
-		if strings.Contains(err.Error(), "Duplicate entry") {
-			return 0, ErrDomainTaken
-		}
-		return 0, fmt.Errorf("route insert: %w", err)
-	}
-	routeID, _ := res.LastInsertId()
-	// Claim the slot conditionally, in the same transaction as the insert.
-	// Placement read current_routes < max_routes outside any lock, so two
-	// concurrent creates could both see the same last free slot and both bump
-	// the counter past max_routes. The WHERE clause makes the claim atomic:
-	// exactly one of them updates a row, the other is told to retry.
-	claim, err := tx.ExecContext(ctx,
-		"UPDATE caddy_nodes SET current_routes = current_routes + 1 WHERE id = ? AND current_routes < max_routes",
-		nodeID)
-	if err != nil {
-		return 0, fmt.Errorf("node counter bump: %w", err)
-	}
-	if n, _ := claim.RowsAffected(); n == 0 {
-		// Single-mode placement can simply move to another node with room.
-		// Not attempted for a tunnel route (the peer was validated against the
-		// originally placed node) or for fan-out modes (the peer set was chosen
-		// as a whole), where the caller retries instead.
-		alt, ok := int64(0), false
-		if groupMode == "single" && in.ViaWGPeerID == 0 {
-			alt, ok = claimNodeWithCapacity(ctx, tx, nodeGroupID, nodeID)
-		}
-		if !ok {
-			return 0, ErrNodeAtCapacity
-		}
-		if _, err := tx.ExecContext(ctx,
-			"UPDATE routes SET caddy_node_id = ? WHERE id = ?", alt, routeID); err != nil {
-			return 0, fmt.Errorf("re-place route: %w", err)
-		}
-		s.Logger.Warn("primary node filled up between placement and claim; re-placed",
-			"route_id", routeID, "domain", domain, "from_node", nodeID, "to_node", alt)
-		nodeID = alt
-		allNodes = []int64{alt}
-	}
-	// Fan-out modes: record every target node in the assignments join table.
-	// active_active deploys to all peers; failover deploys to primary + one
-	// warm standby so the standby has the route ready when it's promoted.
-	// assignedNodes is the set whose slots we actually hold, which is what the
-	// post-commit pushes must target: a peer skipped for capacity has no
-	// assignment row, so pushing to it would be a no-op round trip.
-	assignedNodes := []int64{nodeID}
-	if groupMode != "single" && len(allNodes) > 1 {
-		for _, n := range allNodes {
-			if n != nodeID {
-				// Same atomic claim as the primary. A peer that filled up since
-				// placement is skipped rather than oversubscribed: the route
-				// still lands on the peers that had room, and the assignment is
-				// only recorded for nodes whose slot we actually hold.
-				peerClaim, err := tx.ExecContext(ctx,
-					"UPDATE caddy_nodes SET current_routes = current_routes + 1 WHERE id = ? AND current_routes < max_routes",
-					n)
-				if err != nil {
-					return 0, fmt.Errorf("peer counter bump: %w", err)
-				}
-				if got, _ := peerClaim.RowsAffected(); got == 0 {
-					s.Logger.Warn("fan-out peer at capacity, skipped",
-						"route_id", routeID, "node_id", n, "domain", domain)
-					continue
-				}
-			}
-			if _, err := tx.ExecContext(ctx,
-				store.InsertOrIgnore()+" INTO route_node_assignments (route_id, node_id) VALUES (?, ?)",
-				routeID, n); err != nil {
-				return 0, fmt.Errorf("fan-out assign: %w", err)
-			}
-			if n != nodeID {
-				assignedNodes = append(assignedNodes, n)
-			}
-		}
-	}
-	if err := tx.Commit(); err != nil {
-		return 0, err
-	}
-
-	// Best-effort lifecycle. Failures are recorded in the row, not returned.
-	// Bound it: advanceRoute does a DNS lookup + Caddy push and must not pile
-	// up holding a DB connection unbounded under burst route creation.
-	go func() {
-		defer recoverBg(s.Logger, "advanceRoute")
-		ctx, cancel := context.WithTimeout(s.BackgroundCtx(), 45*time.Second)
-		defer cancel()
-		s.advanceRoute(ctx, routeID)
-	}()
-	if groupMode != "single" {
-		for _, n := range assignedNodes {
-			if n == nodeID {
-				continue // the anchor is pushed by advanceRoute
-			}
-			s.schedulePush(n)
-		}
-	}
-	return routeID, nil
-}
-
-// VerifyDNS re-runs the DNS check for an existing route and re-pushes
-// to Caddy if it transitions to dns_ok.
-func (s *Service) VerifyDNS(ctx context.Context, clientID, routeID int64) error {
-	var ownerClient int64
-	if err := s.DB.QueryRowContext(ctx,
-		`SELECT sv.client_id FROM routes r JOIN services sv ON sv.id = r.service_id WHERE r.id = ?`,
-		routeID,
-	).Scan(&ownerClient); err != nil {
-		return err
-	}
-	if clientID != 0 && ownerClient != clientID {
-		return ErrServiceNotYours
-	}
-	s.advanceRoute(ctx, routeID)
-	return nil
-}
-
-// ErrAlreadyVerified: the route's domain is already proven; nothing to do.
-var ErrAlreadyVerified = errors.New("domain already verified")
-
-// ErrVerifyTokenMissing: TXT record at _hpg-verify.<domain> did not contain the
-// route's token (owner hasn't published it yet, or it's wrong).
-var ErrVerifyTokenMissing = errors.New("verification TXT record not found")
-
-// VerifyDomainToken proves the caller controls the route's domain by looking up
-// a DNS TXT record at _hpg-verify.<domain> and matching it against the route's
-// verify_token. On success it marks the route verified, evicts any STALE
-// UNVERIFIED squatter routes on the same domain+path owned by a different tenant
-// (anti-squat takeover), then advances the route. Returns the token+FQDN so the
-// caller can re-surface instructions on failure.
-func (s *Service) VerifyDomainToken(ctx context.Context, clientID, routeID int64) (token, recordName string, err error) {
-	var (
-		ownerClient int64
-		domain      string
-		pathPrefix  sql.NullString
-		verified    int
-		tok         sql.NullString
-		aliases     sql.NullString
-	)
-	if err = s.DB.QueryRowContext(ctx,
-		`SELECT sv.client_id, r.domain, r.path_prefix, r.domain_verified, r.verify_token,
-		        COALESCE(r.aliases,'')
-		 FROM routes r JOIN services sv ON sv.id = r.service_id WHERE r.id = ?`,
-		routeID,
-	).Scan(&ownerClient, &domain, &pathPrefix, &verified, &tok, &aliases); err != nil {
-		return "", "", err
-	}
-	if clientID != 0 && ownerClient != clientID {
-		return "", "", ErrServiceNotYours
-	}
-	token = tok.String
-	recordName = "_hpg-verify." + domain
-	// Aliases are proven independently of the primary domain, so sweep them on
-	// every attempt - including one on an already-verified route.
-	s.verifyAliases(ctx, routeID, aliases.String, token)
-	if verified == 1 {
-		return token, recordName, ErrAlreadyVerified
-	}
-	if token == "" {
-		// No token on the row (shouldn't happen for self-service creates); refuse
-		// rather than silently verifying.
-		return token, recordName, ErrVerifyTokenMissing
-	}
-	if !dns.TXTContains(ctx, recordName, token) {
-		return token, recordName, ErrVerifyTokenMissing
-	}
-
-	// Anti-squat takeover: now that this tenant PROVED ownership, remove any
-	// still-unverified route rows for the same domain+path held by a DIFFERENT
-	// tenant so their stale first-come claim can't keep blocking the real owner.
-	// Only unverified rows are evicted - a verified conflicting claim is left
-	// intact (that owner also proved control; a genuine dispute is out of scope).
-	rows, derr := s.DB.QueryContext(ctx,
-		`SELECT id, caddy_node_id FROM routes
-		 WHERE domain = ? AND COALESCE(path_prefix,'') = COALESCE(?, '')
-		   AND id <> ? AND domain_verified = 0`,
-		domain, pathPrefix, routeID)
-	if derr == nil {
-		type stale struct{ id, node int64 }
-		var stales []stale
-		for rows.Next() {
-			var st stale
-			var node sql.NullInt64
-			if rows.Scan(&st.id, &node) == nil {
-				st.node = node.Int64
-				stales = append(stales, st)
-			}
-		}
-		rows.Close()
-		for _, st := range stales {
-			if _, e := s.DB.ExecContext(ctx, "DELETE FROM routes WHERE id = ? AND domain_verified = 0", st.id); e == nil {
-				s.Logger.Warn("anti-squat: evicted unverified route", "evicted_id", st.id, "domain", domain, "winner_id", routeID)
-				if st.node != 0 {
-					_, _ = s.DB.ExecContext(ctx,
-						"UPDATE caddy_nodes SET current_routes = GREATEST(current_routes - 1, 0) WHERE id = ?", st.node)
-					// Row already gone: drop it from the squatter's node config.
-					nodeID := st.node
-					evID := st.id
-					go func() {
-						defer recoverBg(s.Logger, "antiSquat.removeRoute")
-						c, cancel := context.WithTimeout(s.BackgroundCtx(), 30*time.Second)
-						defer cancel()
-						_ = s.pushRouteIncremental(c, nodeID, evID, routeRemove)
-					}()
-				}
-			}
-		}
-	}
-
-	if _, err = s.DB.ExecContext(ctx,
-		"UPDATE routes SET domain_verified = 1, last_error = NULL, updated_at = NOW() WHERE id = ?",
-		routeID); err != nil {
-		return token, recordName, err
-	}
-	s.advanceRoute(ctx, routeID)
-	return token, recordName, nil
-}
-
-// verifyAliases proves each alias with the same TXT nonce at _hpg-verify.<alias>
-// and records the proven subset in routes.aliases_verified. An alias that is not
-// proven is never emitted as a host matcher and never gets a certificate, so an
-// operator cannot bolt a victim hostname onto a route they already own.
-// Returns the proven subset after the sweep.
-func (s *Service) verifyAliases(ctx context.Context, routeID int64, aliases, token string) []string {
-	list := splitHostList(aliases)
-	if token == "" || len(list) == 0 {
-		return nil
-	}
-	var prev sql.NullString
-	if err := s.DB.QueryRowContext(ctx,
-		"SELECT COALESCE(aliases_verified,'') FROM routes WHERE id = ?", routeID).Scan(&prev); err != nil {
-		return nil
-	}
-	proven := map[string]bool{}
-	for _, a := range splitHostList(prev.String) {
-		proven[a] = true
-	}
-	out := []string{}
-	changed := false
-	for _, a := range list {
-		if !proven[a] {
-			if !dns.TXTContains(ctx, "_hpg-verify."+a, token) {
-				continue
-			}
-			changed = true
-		}
-		out = append(out, a)
-	}
-	// Also collapses a stale entry for a removed alias.
-	joined := strings.Join(out, ",")
-	if !changed && joined == strings.Join(splitHostList(prev.String), ",") {
-		return out
-	}
-	if _, err := s.DB.ExecContext(ctx,
-		"UPDATE routes SET aliases_verified = ? WHERE id = ?", joined, routeID); err != nil {
-		s.Logger.Warn("alias verification: persist", "route_id", routeID, "err", err)
-		return splitHostList(prev.String)
-	}
-	var nodeID sql.NullInt64
-	if err := s.DB.QueryRowContext(ctx,
-		"SELECT caddy_node_id FROM routes WHERE id = ?", routeID).Scan(&nodeID); err == nil && nodeID.Int64 > 0 {
-		s.SchedulePush(nodeID.Int64)
-	}
-	return out
-}
-
-// RecheckPendingAliases re-runs the TXT proof for every route that still has
-// unproven aliases. Migration 00138 dropped the 00136 backfill, so an owner
-// whose _hpg-verify record is already published recovers here with no manual
-// step; a legacy claim that fully re-proves is closed out automatically.
-func (s *Service) RecheckPendingAliases(ctx context.Context) {
-	if s.DB == nil {
-		return
-	}
-	rows, err := s.DB.QueryContext(ctx,
-		`SELECT id, COALESCE(aliases,''), COALESCE(aliases_verified,''), COALESCE(verify_token,'')
-		   FROM routes
-		  WHERE aliases IS NOT NULL AND aliases <> '' AND status <> 'disabled'
-		  ORDER BY id ASC LIMIT 500`)
-	if err != nil {
-		s.Logger.Warn("alias recheck: list", "err", err)
-		return
-	}
-	type pending struct {
-		id             int64
-		aliases, token string
-	}
-	var todo []pending
-	for rows.Next() {
-		var p pending
-		var verified string
-		if rows.Scan(&p.id, &p.aliases, &verified, &p.token) != nil {
-			continue
-		}
-		if p.token == "" || len(unprovenHosts(p.aliases, verified)) == 0 {
-			continue
-		}
-		todo = append(todo, p)
-	}
-	rows.Close()
-	for _, p := range todo {
-		if ctx.Err() != nil {
-			return
-		}
-		proven := s.verifyAliases(ctx, p.id, p.aliases, p.token)
-		if len(unprovenHosts(p.aliases, strings.Join(proven, ","))) > 0 {
-			continue
-		}
-		_, _ = s.DB.ExecContext(ctx,
-			`UPDATE route_alias_legacy_claims SET status='proven', resolved_at=NOW()
-			  WHERE route_id = ? AND status = 'pending'`, p.id)
-	}
-}
-
-// unprovenHosts returns the entries of aliases that are absent from verified.
-func unprovenHosts(aliases, verified string) []string {
-	ok := map[string]bool{}
-	for _, v := range splitHostList(verified) {
-		ok[v] = true
-	}
-	out := []string{}
-	for _, a := range splitHostList(aliases) {
-		if !ok[a] {
-			out = append(out, a)
-		}
-	}
-	return out
-}
-
-// Delete removes the route, decrements the node counter, and rebuilds the
-// node config so Caddy stops serving the domain.
-func (s *Service) Delete(ctx context.Context, clientID, routeID int64) error {
-	var ownerClient, nodeID int64
-	if err := s.DB.QueryRowContext(ctx,
-		`SELECT sv.client_id, r.caddy_node_id FROM routes r
-		 JOIN services sv ON sv.id = r.service_id WHERE r.id = ?`,
-		routeID,
-	).Scan(&ownerClient, &nodeID); err != nil {
-		return err
-	}
-	if clientID != 0 && ownerClient != clientID {
-		return ErrServiceNotYours
-	}
-
-	// Collect all fan-out nodes before the transaction so we can decrement
-	// every node that holds a copy of this route.
-	fanOutNodes, err := s.fanOutNodes(ctx, routeID, nodeID)
-	if err != nil {
-		return err
-	}
-
-	tx, err := s.DB.BeginTx(ctx, nil)
-	if err != nil {
-		return err
-	}
-	defer tx.Rollback() //nolint:errcheck
-	if _, err := tx.ExecContext(ctx, "DELETE FROM routes WHERE id = ?", routeID); err != nil {
-		return err
-	}
-	// Decrement primary node.
-	if _, err := tx.ExecContext(ctx,
-		"UPDATE caddy_nodes SET current_routes = GREATEST(current_routes - 1, 0) WHERE id = ?", nodeID); err != nil {
-		return err
-	}
-	// Decrement fan-out peers and clean up assignments table.
-	if len(fanOutNodes) > 0 {
-		for _, peerID := range fanOutNodes {
-			if _, err := tx.ExecContext(ctx,
-				"UPDATE caddy_nodes SET current_routes = GREATEST(current_routes - 1, 0) WHERE id = ?", peerID); err != nil {
-				return err
-			}
-		}
-		if _, err := tx.ExecContext(ctx,
-			"DELETE FROM route_node_assignments WHERE route_id = ?", routeID); err != nil {
-			return err
-		}
-	}
-	if err := tx.Commit(); err != nil {
-		return err
-	}
-
-	allNodes := append([]int64{nodeID}, fanOutNodes...)
-	go func() {
-		defer recoverBg(s.Logger, "pushRouteIncremental.remove")
-		ctx, cancel := context.WithTimeout(s.BackgroundCtx(), 30*time.Second)
-		defer cancel()
-		// Row is already deleted: remove the route from every node it was on.
-		for _, nid := range allNodes {
-			_ = s.pushRouteIncremental(ctx, nid, routeID, routeRemove)
-		}
-	}()
-	return nil
-}
-
-// fanOutNodes returns node IDs in route_node_assignments for routeID,
-// excluding the primary node (already tracked via caddy_node_id).
-func (s *Service) fanOutNodes(ctx context.Context, routeID, primaryNodeID int64) ([]int64, error) {
-	rows, err := s.DB.QueryContext(ctx,
-		"SELECT node_id FROM route_node_assignments WHERE route_id = ? AND node_id != ?",
-		routeID, primaryNodeID)
-	if err != nil {
-		return nil, err
-	}
-	defer rows.Close()
-	var ids []int64
-	for rows.Next() {
-		var id int64
-		if err := rows.Scan(&id); err != nil {
-			return nil, err
-		}
-		ids = append(ids, id)
-	}
-	return ids, rows.Err()
-}
-
-// advanceRoute: DNS check → status update → push if eligible.
-func (s *Service) advanceRoute(ctx context.Context, routeID int64) {
-	var (
-		nodeID       int64
-		domain       string
-		nodeHostname sql.NullString
-		nodeIP       sql.NullString
-		verified     int
-	)
-	if err := s.DB.QueryRowContext(ctx,
-		`SELECT r.caddy_node_id, r.domain, n.public_hostname, n.public_ip, r.domain_verified
-		 FROM routes r JOIN caddy_nodes n ON n.id = r.caddy_node_id WHERE r.id = ?`,
-		routeID,
-	).Scan(&nodeID, &domain, &nodeHostname, &nodeIP, &verified); err != nil {
-		s.Logger.Error("advance: route lookup", "id", routeID, "err", err)
-		return
-	}
-
-	// Domain-ownership gate: an unverified route must not advance past pending_dns
-	// (no serving, no cert). Pin it to pending_dns so Reconcile keeps retrying and
-	// the owner sees the "verify domain" state until the TXT proof clears it.
-	if verified == 0 {
-		_, _ = s.DB.ExecContext(ctx,
-			"UPDATE routes SET status='pending_dns', last_error='domain ownership not verified', updated_at=NOW() WHERE id=?",
-			routeID)
-		s.Logger.Info("route: domain unverified, holding", "id", routeID, "domain", domain)
-		return
-	}
-
-	if err := dns.Check(ctx, domain, nodeHostname.String, nodeIP.String); err != nil {
-		_, _ = s.DB.ExecContext(ctx,
-			"UPDATE routes SET status='pending_dns', last_error=?, dns_checked_at=NOW(), updated_at=NOW() WHERE id=?",
-			truncErr(err), routeID)
-		s.Logger.Info("route: dns pending", "id", routeID, "domain", domain, "err", err)
-		// Push anyway: Caddy serves HTTP-01 challenge on :80 once DNS catches up.
-		// For initial MVP we wait for DNS to be correct before pushing.
-		return
-	}
-	_, _ = s.DB.ExecContext(ctx,
-		"UPDATE routes SET status='dns_ok', last_error=NULL, dns_checked_at=NOW(), updated_at=NOW() WHERE id=?", routeID)
-
-	// Incremental single-route push (covers Create-primary, VerifyDNS, Reconcile,
-	// which all funnel through advanceRoute); falls back to full /load on error.
-	if err := s.pushRouteIncremental(ctx, nodeID, routeID, routeUpsert); err != nil {
-		_, _ = s.DB.ExecContext(ctx,
-			"UPDATE routes SET status='failed', last_error=? WHERE id=?", truncErr(err), routeID)
-		if s.Webhooks != nil {
-			s.Webhooks.Emit(ctx, "route.failed", map[string]any{
-				"route_id": routeID, "domain": domain, "node_id": nodeID,
-				"error": truncErr(err),
-			})
-		}
-		return
-	}
-	// On first activation set ssl_issued_at; on subsequent activations
-	// (renewal) refresh it so the certs page shows recent activity.
-	_, _ = s.DB.ExecContext(ctx,
-		`UPDATE routes
-		   SET status='active', last_error=NULL,
-		       ssl_issued_at = CASE WHEN ssl_enabled = 1 THEN NOW() ELSE ssl_issued_at END
-		 WHERE id=?`, routeID)
-	// Fan-out peers (active_active/failover) got routes:0 at create time while
-	// the route was still pending; re-push them now that it is active.
-	if peers, perr := s.fanOutNodes(ctx, routeID, nodeID); perr == nil {
-		for _, pid := range peers {
-			s.schedulePush(pid)
-		}
-	}
-	if s.Webhooks != nil {
-		s.Webhooks.Emit(ctx, "route.active", map[string]any{
-			"route_id": routeID, "domain": domain, "node_id": nodeID,
-		})
-	}
-}
-
-// HealthProbe sweeps every enabled node in parallel, GETs its /config/
-// endpoint, and updates health_status + last_seen_at. Errors are logged,
-// not returned. Bounded concurrency keeps slow nodes from delaying the
-// rest. Designed to be called every ~30s from a background ticker.
-const healthProbeWorkers = 8
-
-// pushWorkers bounds concurrent full /load pushes during fleet-wide sweeps
-// (boot push, drift resync). Without this a single slow node serialized the
-// whole sweep at N x clientTimeout; now one slow node only ties up one worker.
-const pushWorkers = 4
-
-// reconcileWorkers bounds concurrent per-route/per-node work in the reconcile
-// sweeps so one slow node can't stall the whole sweep (it previously ran serial).
-const reconcileWorkers = 4
-
-// pushNodesConcurrent fans out pushNodeConfig across ids with bounded
-// concurrency and a per-node timeout, so one slow/half-open node cannot stall
-// the entire sweep. Errors are logged, not returned. Respects ctx cancel.
-func (s *Service) pushNodesConcurrent(ctx context.Context, ids []int64, label string) {
-	sem := make(chan struct{}, pushWorkers)
-	var wg sync.WaitGroup
-	for _, id := range ids {
-		select {
-		case <-ctx.Done():
-			wg.Wait()
-			return
-		case sem <- struct{}{}:
-		}
-		wg.Add(1)
-		go func(id int64) {
-			defer recoverBg(s.Logger, "pushAll")
-			defer wg.Done()
-			defer func() { <-sem }()
-			pushCtx, cancel := context.WithTimeout(ctx, 30*time.Second)
-			defer cancel()
-			if err := s.pushNodeConfig(pushCtx, id); err != nil {
-				s.Logger.Warn(label+" failed", "node_id", id, "err", err)
-			} else {
-				s.Logger.Info(label+" ok", "node_id", id)
-			}
-		}(id)
-	}
-	wg.Wait()
-}
-
-func (s *Service) HealthProbe(ctx context.Context) {
-	if s.DB == nil {
-		return
-	}
-	rows, err := s.DB.QueryContext(ctx,
-		"SELECT id, api_url FROM caddy_nodes WHERE is_enabled = 1")
-	if err != nil {
-		s.Logger.Warn("health: list nodes", "err", err)
-		return
-	}
-	type nodeProbe struct {
-		id     int64
-		apiURL string
-	}
-	var probes []nodeProbe
-	for rows.Next() {
-		var p nodeProbe
-		if err := rows.Scan(&p.id, &p.apiURL); err == nil {
-			probes = append(probes, p)
-		}
-	}
-	rows.Close()
-
-	sem := make(chan struct{}, healthProbeWorkers)
-	var wg sync.WaitGroup
-	for _, p := range probes {
-		p := p
-		sem <- struct{}{}
-		wg.Add(1)
-		go func() {
-			defer recoverBg(s.Logger, "healthProbe")
-			defer wg.Done()
-			defer func() { <-sem }()
-			status := "down"
-			client := caddyapi.New(p.apiURL)
-			probeCtx, cancel := context.WithTimeout(ctx, 3*time.Second)
-			probeStart := time.Now()
-			_, probeErr := client.GetRaw(probeCtx, "/config/")
-			rttMs := int(time.Since(probeStart).Milliseconds())
-			if probeErr == nil {
-				status = "healthy"
-			}
-			cancel()
-			_, _ = s.DB.ExecContext(ctx,
-				"UPDATE caddy_nodes SET health_status = ?, last_seen_at = NOW() WHERE id = ?",
-				status, p.id)
-			// Only a successful probe reflects real RTT - a failed/timed-out
-			// probe measures the timeout, not the network.
-			if probeErr == nil {
-				s.recordRTT(ctx, p.id, rttMs, probeStart)
-			}
-
-			// Auto-resync: node came back online (was down/unknown, now healthy).
-			// Caddy may have lost its config on restart, so re-push from DB.
-			if status == "healthy" && s.markHealthAndChanged(p.id, status) {
-				go func(id int64) {
-					defer recoverBg(s.Logger, "autoResync")
-					pushCtx, c := context.WithTimeout(s.BackgroundCtx(), 30*time.Second)
-					defer c()
-					if err := s.pushNodeConfig(pushCtx, id); err != nil {
-						s.Logger.Warn("auto-resync on recovery failed", "node_id", id, "err", err)
-					} else {
-						s.Logger.Info("auto-resync on recovery ok", "node_id", id)
-					}
-				}(p.id)
-			} else {
-				s.markHealth(p.id, status)
-			}
-		}()
-	}
-	wg.Wait()
-	s.pruneRTTSamples(ctx)
-}
-
-// markHealthAndChanged records new status and returns true iff this is a
-// recovery transition (previous status was "down" or absent and new is healthy).
-// First-observation healthy returns false (handled by boot-time PushAll).
-func (s *Service) markHealthAndChanged(id int64, status string) bool {
-	s.healthMu.Lock()
-	defer s.healthMu.Unlock()
-	if s.lastHealth == nil {
-		s.lastHealth = map[int64]string{}
-	}
-	prev, seen := s.lastHealth[id]
-	s.lastHealth[id] = status
-	return seen && prev == "down" && status == "healthy"
-}
-
-func (s *Service) markHealth(id int64, status string) {
-	s.healthMu.Lock()
-	defer s.healthMu.Unlock()
-	if s.lastHealth == nil {
-		s.lastHealth = map[int64]string{}
-	}
-	s.lastHealth[id] = status
-}
-
-// PushAll pushes the current DB-derived config to every enabled node.
-// Used on panel boot so a cold-started Caddy (lost autosave, fresh container)
-// gets repopulated immediately instead of waiting up to 5min for ReconcileDrift.
-func (s *Service) PushAll(ctx context.Context) {
-	if s.DB == nil {
-		return
-	}
-	rows, err := s.DB.QueryContext(ctx,
-		"SELECT id FROM caddy_nodes WHERE is_enabled = 1")
-	if err != nil {
-		s.Logger.Warn("boot push: list nodes", "err", err)
-		return
-	}
-	var ids []int64
-	for rows.Next() {
-		var id int64
-		if err := rows.Scan(&id); err == nil {
-			ids = append(ids, id)
-		}
-	}
-	rows.Close()
-	s.pushNodesConcurrent(ctx, ids, "boot push")
-	if s.AfterPush != nil {
-		s.AfterPush(ctx)
-	}
-}
-
-// AutoFailover migrates routes off Caddy nodes that have been "down"
-// for more than the grace window onto the lowest-loaded healthy peer
-// in the same node_group. Routes bound to a WG tunnel (via_wg_peer_id)
-// are SKIPPED - their wg-tun0 only exists on the failed node, so
-// moving them silently would yield 502s with no clear cause. Those
-// routes need explicit operator intervention (HA tunnel mode or
-// re-issue tunnel on a new node).
-//
-// Called by a leader-only ticker on a 2-min cadence; designed to be
-// idempotent (no-op when no node is down or no peer has capacity).
-const failoverGraceMinutes = 5
-
-func (s *Service) AutoFailover(ctx context.Context) {
-	if s.DB == nil {
-		return
-	}
-	rows, err := s.DB.QueryContext(ctx,
-		`SELECT r.id, r.domain, r.caddy_node_id, n.node_group_id, r.via_wg_peer_id, sv.client_id
-		 FROM routes r
-		 JOIN caddy_nodes n ON n.id = r.caddy_node_id
-		 JOIN services sv ON sv.id = r.service_id
-		 WHERE n.is_enabled = 1
-		   AND n.health_status = 'down'
-		   AND n.last_seen_at < `+store.DateSubParam("MINUTE")+`
-		   AND r.status IN ('active','dns_ok','pending_ssl')
-		 ORDER BY r.id ASC LIMIT 500`, failoverGraceMinutes)
-	if err != nil {
-		s.Logger.Warn("autofailover: list candidates", "err", err)
-		return
-	}
-	type candidate struct {
-		id         int64
-		domain     string
-		fromNodeID int64
-		groupID    int64
-		viaPeerID  sql.NullInt64
-		clientID   int64
-	}
-	var cands []candidate
-	for rows.Next() {
-		var c candidate
-		if err := rows.Scan(&c.id, &c.domain, &c.fromNodeID, &c.groupID, &c.viaPeerID, &c.clientID); err == nil {
-			cands = append(cands, c)
-		}
-	}
-	rows.Close()
-	if len(cands) == 0 {
-		return
-	}
-
-	// Group-level dest cache so we don't re-pick the same dest for every
-	// route. Also avoids the worst-case N picks across the same group.
-	destByGroup := map[int64]int64{}
-	movedByDest := map[int64]int{}
-	for _, c := range cands {
-		select {
-		case <-ctx.Done():
-			return
-		default:
-		}
-		if c.viaPeerID.Valid {
-			// Tunneled routes can't follow - their wg-tun0 is on the
-			// dead node. Surface via webhook so the operator can act.
-			s.Logger.Warn("autofailover: route bound to tunnel, skipping",
-				"route_id", c.id, "domain", c.domain, "from_node", c.fromNodeID)
-			if s.Webhooks != nil {
-				s.Webhooks.Emit(ctx, "route.failover.skipped", map[string]any{
-					"route_id": c.id, "domain": c.domain, "node_id": c.fromNodeID,
-					"reason": "bound to WG tunnel which lives on the failed node",
-				})
-			}
-			if s.Notifier != nil {
-				s.Notifier.Notify(ctx, c.clientID,
-					"[Hostyt] "+c.domain+" cannot auto-failover",
-					"Your route "+c.domain+" lives on a node that went down, and is bound to a WG tunnel "+
-						"only available on that node. Manual action required: either re-create the tunnel on "+
-						"a different node, or enable HA tunnel mode. The site will stay offline until then.")
-			}
-			continue
-		}
-
-		dest, ok := destByGroup[c.groupID]
-		if !ok {
-			var d sql.NullInt64
-			err := s.DB.QueryRowContext(ctx,
-				`SELECT id FROM caddy_nodes
-				 WHERE node_group_id = ? AND id <> ?
-				   AND is_enabled = 1 AND approved_at IS NOT NULL
-				   AND health_status = 'healthy'
-				   AND current_routes < max_routes
-				 ORDER BY (current_routes / GREATEST(max_routes,1)) ASC, priority DESC, id ASC
-				 LIMIT 1`, c.groupID, c.fromNodeID).Scan(&d)
-			if err != nil || !d.Valid {
-				s.Logger.Warn("autofailover: no healthy peer in group", "group_id", c.groupID, "route_id", c.id)
-				continue
-			}
-			dest = d.Int64
-			destByGroup[c.groupID] = dest
-		}
-
-		_, err := s.DB.ExecContext(ctx,
-			`UPDATE routes SET caddy_node_id = ?, updated_at = NOW() WHERE id = ?`,
-			dest, c.id)
-		if err != nil {
-			s.Logger.Warn("autofailover: route update", "route_id", c.id, "err", err)
-			continue
-		}
-		movedByDest[dest]++
-		s.Logger.Info("autofailover moved route", "route_id", c.id, "domain", c.domain,
-			"from_node", c.fromNodeID, "to_node", dest)
-		if s.Webhooks != nil {
-			s.Webhooks.Emit(ctx, "route.failover", map[string]any{
-				"route_id": c.id, "domain": c.domain,
-				"from_node": c.fromNodeID, "to_node": dest,
-			})
-		}
-		if s.Notifier != nil {
-			s.Notifier.Notify(ctx, c.clientID,
-				"[Hostyt] "+c.domain+" moved to a backup node",
-				"The Caddy node serving your route "+c.domain+" went down. We automatically "+
-					"moved your route to a healthy peer in the same group. Your site should be "+
-					"reachable again within a minute. No action required.")
-		}
-	}
-
-	// One push per destination node, not per route - saves N-1 /load calls.
-	// Bounded concurrency so a slow dest node doesn't stall the others.
-	gf, gfctx := errgroup.WithContext(ctx)
-	gf.SetLimit(reconcileWorkers)
-	for destID, n := range movedByDest {
-		destID, n := destID, n
-		gf.Go(func() error {
-			_, _ = s.DB.ExecContext(gfctx,
-				`UPDATE caddy_nodes SET current_routes = current_routes + ? WHERE id = ?`, n, destID)
-			if err := s.pushNodeConfig(gfctx, destID); err != nil {
-				s.Logger.Warn("autofailover: push to new home failed", "node_id", destID, "err", err)
-			}
-			return nil
-		})
-	}
-	_ = gf.Wait()
-	// Best-effort: also bump down-node counters to reflect moves.
-	for _, c := range cands {
-		if c.viaPeerID.Valid {
-			continue
-		}
-		_, _ = s.DB.ExecContext(ctx,
-			`UPDATE caddy_nodes SET current_routes = GREATEST(0, current_routes - 1) WHERE id = ?`, c.fromNodeID)
-	}
-}
-
-// Reconcile picks up routes stuck in non-terminal states beyond a grace
-// window and re-runs advanceRoute on them. Idempotent - if the route is
-// already healthy it's a no-op. Called on a slow ticker (60s).
-func (s *Service) Reconcile(ctx context.Context) {
-	if s.DB == nil {
-		return
-	}
-	rows, err := s.DB.QueryContext(ctx,
-		`SELECT id FROM routes
-		 WHERE status IN ('pending_dns','dns_ok','pending_ssl','failed')
-		   AND updated_at < `+store.DateSub(1, "MINUTE")+`
-		 ORDER BY updated_at ASC LIMIT 100`)
-	if err != nil {
-		s.Logger.Warn("reconcile: list stuck routes", "err", err)
-		return
-	}
-	var ids []int64
-	for rows.Next() {
-		var id int64
-		if err := rows.Scan(&id); err == nil {
-			ids = append(ids, id)
-		}
-	}
-	rows.Close()
-	if len(ids) == 0 {
-		return
-	}
-	s.Logger.Info("reconcile: retry stuck routes", "n", len(ids))
-	// Bounded concurrency: advanceRoute does a DNS check + push per route;
-	// running 4 at once is safe because pushRouteIncremental/pushNodeConfig take
-	// the per-node lock (same-node ops serialize, cross-node parallelize).
-	g, gctx := errgroup.WithContext(ctx)
-	g.SetLimit(reconcileWorkers)
-	for _, id := range ids {
-		id := id
-		g.Go(func() error {
-			if gctx.Err() != nil {
-				return nil
-			}
-			s.advanceRoute(gctx, id)
-			return nil
-		})
-	}
-	_ = g.Wait()
-}
-
-// ReconcileDrift walks every enabled node, fetches its current Caddy
-// route list, computes a fingerprint, and compares to the DB-derived
-// expected fingerprint. Mismatch → trigger a full Resync. Cheap when
-// routes are stable.
-func (s *Service) ReconcileDrift(ctx context.Context) {
-	if s.DB == nil {
-		return
-	}
-	rows, err := s.DB.QueryContext(ctx,
-		"SELECT id, api_url FROM caddy_nodes WHERE is_enabled = 1")
-	if err != nil {
-		return
-	}
-	type node struct {
-		id     int64
-		apiURL string
-	}
-	var nodes []node
-	for rows.Next() {
-		var n node
-		if err := rows.Scan(&n.id, &n.apiURL); err == nil {
-			nodes = append(nodes, n)
-		}
-	}
-	rows.Close()
-	// Probe + resync each node concurrently: the 5s GET plus a possible full
-	// /load is otherwise serial, so one slow node delayed the whole sweep.
-	sem := make(chan struct{}, pushWorkers)
-	var wg sync.WaitGroup
-	for _, n := range nodes {
-		select {
-		case <-ctx.Done():
-			wg.Wait()
-			return
-		case sem <- struct{}{}:
-		}
-		wg.Add(1)
-		go func(n node) {
-			defer recoverBg(s.Logger, "reconcileDrift")
-			defer wg.Done()
-			defer func() { <-sem }()
-			expected, err := s.expectedNodeHash(ctx, n.id)
-			if err != nil {
-				return
-			}
-			client := caddyapi.New(n.apiURL)
-			probeCtx, cancel := context.WithTimeout(ctx, 5*time.Second)
-			actualRaw, err := client.GetRaw(probeCtx, "/config/apps/http/servers/srv0/routes")
-			cancel()
-			if err != nil {
-				return
-			}
-			// Canonicalise before hashing: Caddy may reformat the GET response
-			// (map key order, whitespace) so a raw hash flaps even when the
-			// route set is identical, triggering an infinite resync loop.
-			// Unmarshal+re-marshal gives stable byte output on both sides.
-			// Drop infra routes (panel_self, hpg_wstunnel_*) the expected hash
-			// never carries - else a panel/WSS node drifts every cycle forever.
-			actual := canonHashBytes(filterVirtualRoutes(actualRaw))
-			if actual == expected {
-				return
-			}
-			s.Logger.Warn("drift detected, re-pushing", "node_id", n.id, "expected", expected[:12], "actual", actual[:12])
-			if s.Metrics != nil {
-				s.Metrics.CaddyDriftResync()
-			}
-			pushCtx, c := context.WithTimeout(ctx, 30*time.Second)
-			defer c()
-			if err := s.pushNodeConfig(pushCtx, n.id); err != nil {
-				s.Logger.Warn("drift resync failed", "node_id", n.id, "err", err)
-			}
-		}(n)
-	}
-	wg.Wait()
-}
-
-// Resync rebuilds the node's Caddy config from DB and POSTs /load.
-// Public wrapper around pushNodeConfig for admin use.
-func (s *Service) Resync(ctx context.Context, nodeID int64) error {
-	err := s.pushNodeConfig(ctx, nodeID)
-	if err == nil && s.AfterPush != nil {
-		s.AfterPush(ctx)
-	}
-	return err
-}
-
-// nodePush is the built, ready-to-Load config for one node plus the data
-// loadNodeConfig needs for fingerprinting.
-type nodePush struct {
-	cfg      map[string]any
-	built    []caddyapi.Route
-	routeIDs []int64
-	apiURL   string
-}
-
-// validTunnelHostname accepts only DNS-hostname / IPv4 characters. Rejects
-// schemes, IPv6 brackets/colons, and any junk so a bad tunnel_endpoint never
-// reaches Caddy JSON (which would fail /load for the whole node).
-func validTunnelHostname(h string) bool {
-	if h == "" || len(h) > 253 {
-		return false
-	}
-	for _, r := range h {
-		if !(r >= 'a' && r <= 'z' || r >= 'A' && r <= 'Z' || r >= '0' && r <= '9' || r == '.' || r == '-') {
-			return false
-		}
-	}
-	return true
-}
-
-// buildNodePush renders the full Caddy config for a node from DB. Read-only;
-// holds no lock.
-func (s *Service) buildNodePush(ctx context.Context, nodeID int64) (*nodePush, error) {
-	built, routeIDs, err := s.buildRoutesForNode(ctx, nodeID)
-	if err != nil {
-		return nil, err
-	}
-	var (
-		apiURL              string
-		transport           sql.NullString
-		wstunnelPort        sql.NullInt64
-		tunnelEndpoint      sql.NullString
-		tunnelEnabled       bool
-		wstHealthy          sql.NullBool
-		wstFresh            sql.NullBool
-		proxyProtoIn        bool
-		proxyProtoAllow     string
-		proxyProtoTimeoutMs int
-	)
-	var nodeHasWAF, nodeHasL4, nodeHasGeoIP, nodeHasRateLimit, nodeHasDNS sql.NullBool
-	if err := s.DB.QueryRowContext(ctx,
-		`SELECT api_url, tunnel_transport, tunnel_wstunnel_port, tunnel_endpoint, tunnel_enabled,
-		        tunnel_wstunnel_healthy,
-		        tunnel_wstunnel_reported_at > `+store.DateSub(3, "MINUTE")+`,
-		        CASE WHEN modules_probed_at IS NOT NULL THEN has_waf       ELSE NULL END,
-		        CASE WHEN modules_probed_at IS NOT NULL THEN has_l4        ELSE NULL END,
-		        CASE WHEN modules_probed_at IS NOT NULL THEN has_geoip     ELSE NULL END,
-		        CASE WHEN modules_probed_at IS NOT NULL THEN has_rate_limit ELSE NULL END,
-		        CASE WHEN modules_probed_at IS NOT NULL THEN has_dns_module ELSE NULL END,
-		        proxy_protocol_in, proxy_protocol_allow, proxy_protocol_timeout_ms
-		   FROM caddy_nodes WHERE id = ?`,
-		nodeID).Scan(&apiURL, &transport, &wstunnelPort, &tunnelEndpoint, &tunnelEnabled,
-		&wstHealthy, &wstFresh,
-		&nodeHasWAF, &nodeHasL4, &nodeHasGeoIP, &nodeHasRateLimit, &nodeHasDNS,
-		&proxyProtoIn, &proxyProtoAllow, &proxyProtoTimeoutMs); err != nil {
-		return nil, err
-	}
-	// If the node has been probed, its per-node capability flags are authoritative.
-	// Unprobed nodes fall back to the global env-configured flags so existing
-	// deployments are not affected before the first probe runs.
-	probedOr := func(probed sql.NullBool, global bool) bool {
-		if probed.Valid {
-			return probed.Bool
-		}
-		return global
-	}
-	// Fail closed: an unscreened stream set must never reach a node.
-	streams, err := s.buildStreamsForNode(ctx, nodeID)
-	if err != nil {
-		return nil, err
-	}
-	branding := s.loadErrorBranding(ctx)
-	for i := range built {
-		built[i].ErrorBranding = branding
-	}
-
-	// Build wstunnel Caddy route when transport is not pure UDP. Fail-closed:
-	// a malformed endpoint host (scheme, IPv6, junk) must NOT be emitted into
-	// Caddy JSON - that would fail the node's /load and break ALL routes.
-	// Also gate on node health: emit only when the node reported a healthy
-	// wstunnel recently, OR has not reported yet (just-enabled). A node that
-	// reports unhealthy/stale gets no route, so we never advertise dead WSS.
-	healthOK := !wstHealthy.Valid || (wstHealthy.Bool && wstFresh.Valid && wstFresh.Bool)
-	var wstunnelRoute *caddyapi.WstunnelRoute
-	if tunnelEnabled && transport.String != "" && transport.String != "udp" && wstunnelPort.Valid &&
-		wstunnelPort.Int64 > 0 && wstunnelPort.Int64 < 65536 && tunnelEndpoint.Valid && healthOK {
-		hostname, _, _ := net.SplitHostPort(tunnelEndpoint.String)
-		if hostname == "" {
-			hostname = tunnelEndpoint.String
-		}
-		if validTunnelHostname(hostname) {
-			wstunnelRoute = &caddyapi.WstunnelRoute{
-				NodeID:   nodeID,
-				Hostname: hostname,
-				Port:     int(wstunnelPort.Int64),
-			}
-		} else {
-			s.Logger.Warn("skipping wstunnel route: invalid tunnel endpoint host",
-				"node_id", nodeID)
-		}
-	}
-
-	mtlsFailOpen := s.loadMTLSFailOpen(ctx)
-	trustCFIP := s.loadTrustCloudflareIP(ctx)
-	cfg := caddyapi.BuildNodeConfig(built, caddyapi.NodeSettings{
-		ACMEEmail:                s.ACMEEmail,
-		ACMEStaging:              s.ACMEStaging,
-		ACMECaURL:                s.ACMECaURL,
-		ACMEEabKID:               s.ACMEEabKID,
-		ACMEEabHMAC:              s.ACMEEabHMAC,
-		AskURL:                   s.AskURL,
-		PanelRoute:               s.panelRoute(),
-		CacheModuleAvailable:     s.CacheModuleAvailable,
-		Layer4ModuleAvailable:    probedOr(nodeHasL4, s.Layer4ModuleAvailable),
-		RateLimitModuleAvailable: probedOr(nodeHasRateLimit, s.RateLimitModuleAvailable),
-		WAFModuleAvailable:       probedOr(nodeHasWAF, s.WAFModuleAvailable),
-		GeoModuleAvailable:       probedOr(nodeHasGeoIP, s.GeoModuleAvailable) && geoip.HasCountryDB(),
-		DNS01ModuleAvailable:     probedOr(nodeHasDNS, s.DNS01ModuleAvailable),
-		WildcardPolicies:         s.buildWildcardPolicies(ctx, nodeID),
-		StreamRoutes:             streams,
-		ErrorBranding:            branding,
-		WstunnelRoute:            wstunnelRoute,
-		AccessLogURL:             s.AccessLogURL,
-		MTLSFailOpen:             mtlsFailOpen,
-		AdminListen:              s.CaddyAdminListen,
-		TrustCloudflareIP:        trustCFIP,
-		CloudflareRanges:         cloudflare.EdgeCIDRs(),
-		ProxyProtocolIn:          proxyProtoIn,
-		ProxyProtocolAllow:       proxyProtoAllow,
-		ProxyProtocolTimeoutMs:   proxyProtoTimeoutMs,
-		ManualCerts:              s.buildManualCertsForNode(ctx, nodeID),
-	})
-	return &nodePush{cfg: cfg, built: built, routeIDs: routeIDs, apiURL: apiURL}, nil
-}
-
-// loadNodeConfig POSTs the full config (/load) and records the per-route drift
-// fingerprint. The caller MUST hold the per-node lock.
-func (s *Service) loadNodeConfig(ctx context.Context, nodeID int64, np *nodePush) error {
-	client := caddyapi.New(np.apiURL)
-	if err := client.Load(ctx, np.cfg); err != nil {
-		s.Logger.Error("caddy push failed", "node_id", nodeID, "err", err)
-		if s.Metrics != nil {
-			s.Metrics.CaddyPushFail()
-		}
-		return err
-	}
-	if s.Metrics != nil {
-		s.Metrics.CaddyPushOK()
-	}
-	pushHash := hashRoutes(np.built)
-	for _, id := range np.routeIDs {
-		_, _ = s.DB.ExecContext(ctx,
-			"UPDATE routes SET last_pushed_at = NOW(), last_pushed_hash = ? WHERE id = ?",
-			pushHash, id)
-	}
-	s.Logger.Info("caddy push ok", "node_id", nodeID, "routes", len(np.built), "hash", pushHash[:12])
-	return nil
-}
-
-// maxPushGenerations bounds the rebuild loop in pushNodeConfig. A node under a
-// constant stream of edits would otherwise never hand the lock back; the edits
-// that lose their turn have already scheduled their own push.
-const maxPushGenerations = 3
-
-// pushNodeConfig loads the node's full config, building it UNDER the per-node
-// lock and re-checking the config generation afterwards.
-//
-// Building outside the lock (as this used to) let an older snapshot win: build
-// at generation 10, block on the lock while another writer pushed generation
-// 11, then /load the generation-10 snapshot and silently revert it. The lock
-// serialized the request to Caddy but said nothing about how fresh the state
-// behind it was. Building under the lock means a push always reflects DB state
-// at least as new as any push that finished before it; the generation re-check
-// then covers a change committed while this push was in flight, so the node
-// converges immediately instead of waiting for the next debounce or drift
-// sweep.
-func (s *Service) pushNodeConfig(ctx context.Context, nodeID int64) error {
-	lock := s.nodeLock(nodeID)
-	lock.Lock()
-	defer lock.Unlock()
-
-	for attempt := 0; attempt < maxPushGenerations; attempt++ {
-		gen := s.currentGen(nodeID)
-		np, err := s.buildNodePush(ctx, nodeID)
-		if err != nil {
-			return err
-		}
-		if err := s.loadNodeConfig(ctx, nodeID, np); err != nil {
-			return err
-		}
-		s.recordApplied(nodeID, gen)
-		if s.currentGen(nodeID) == gen {
-			return nil
-		}
-		// A change landed while we were pushing: rebuild rather than leave the
-		// node on the snapshot we just applied.
-		if ctx.Err() != nil {
-			return ctx.Err()
-		}
-	}
-	if s.Logger != nil {
-		s.Logger.Warn("node config still changing after repeated pushes; leaving convergence to the next scheduled push",
-			"node_id", nodeID, "attempts", maxPushGenerations)
-	}
-	return nil
-}
-
-// pushNodeConfigLocked is the full-/load fallback for callers that ALREADY hold
-// the per-node lock (pushRouteIncremental). Builds under the held lock; the
-// minor cost only applies on the rare incremental-fallback path.
-func (s *Service) pushNodeConfigLocked(ctx context.Context, nodeID int64) error {
-	np, err := s.buildNodePush(ctx, nodeID)
-	if err != nil {
-		return err
-	}
-	return s.loadNodeConfig(ctx, nodeID, np)
-}
-
-type routeOp int
-
-const (
-	routeUpsert routeOp = iota // add if absent, replace if present
-	routeRemove                // delete by @id
-)
-
-// isNotFound reports whether a Caddy client error is a 404 (already gone).
-func isNotFound(err error) bool {
-	return err != nil && strings.Contains(err.Error(), "404")
-}
-
-// buildOneRoute returns the built, emit-ready Route for routeID on nodeID and
-// whether it is eligible to be emitted. It reuses buildRoutesForNode (then
-// filters) so the emitted object is byte-identical to a full /load element -
-// which keeps drift hashing consistent. ok=false means the route is in DB but
-// filtered out (not active, revoked tunnel, disallowed external, undecryptable
-// secret) and should therefore be absent on the node.
-func (s *Service) buildOneRoute(ctx context.Context, nodeID, routeID int64) (caddyapi.Route, bool, error) {
-	built, ids, err := s.buildRoutesForNode(ctx, nodeID)
-	if err != nil {
-		return caddyapi.Route{}, false, err
-	}
-	branding := s.loadErrorBranding(ctx)
-	// Mirror the full /load module gating (probe OR env): the WAF/Geo/RateLimit
-	// availability must use the node's probed capability, not just the env flag,
-	// or an incremental push emits JSON WITHOUT the handler while a full /load
-	// emits it WITH - suppressing the handler and diverging the config hash
-	// (endless resync). This is why geo blocking silently dropped on nodes whose
-	// module is probe-detected but GEOIP_AVAILABLE is unset.
-	probedOr := func(probed sql.NullBool, global bool) bool {
-		if probed.Valid {
-			return probed.Bool
-		}
-		return global
-	}
-	var nHasWAF, nHasGeoIP, nHasRate sql.NullBool
-	_ = s.DB.QueryRowContext(ctx,
-		`SELECT CASE WHEN modules_probed_at IS NOT NULL THEN has_waf        END,
-		        CASE WHEN modules_probed_at IS NOT NULL THEN has_geoip      END,
-		        CASE WHEN modules_probed_at IS NOT NULL THEN has_rate_limit END
-		   FROM caddy_nodes WHERE id = ?`, nodeID).Scan(&nHasWAF, &nHasGeoIP, &nHasRate)
-	for i, id := range ids {
-		if id == routeID {
-			r := built[i]
-			// Match what BuildNodeConfig sets per-route on a full /load so the
-			// emitted JSON is identical (drift consistency).
-			r.ErrorBranding = branding
-			r.CacheModuleAvailable = s.CacheModuleAvailable
-			r.RateLimitModuleAvailable = probedOr(nHasRate, s.RateLimitModuleAvailable)
-			r.WAFModuleAvailable = probedOr(nHasWAF, s.WAFModuleAvailable)
-			r.GeoModuleAvailable = probedOr(nHasGeoIP, s.GeoModuleAvailable) && geoip.HasCountryDB()
-			return r, true, nil
-		}
-	}
-	return caddyapi.Route{}, false, nil
-}
-
-// routeMatchHosts extracts the host strings from a Caddy route object's match[].
-func routeMatchHosts(obj map[string]any) []string {
-	matches, _ := obj["match"].([]any)
-	var out []string
-	for _, m := range matches {
-		mm, _ := m.(map[string]any)
-		hs, _ := mm["host"].([]any)
-		for _, h := range hs {
-			if str, ok := h.(string); ok {
-				out = append(out, str)
-			}
-		}
-	}
-	return out
-}
-
-// routePresenceAndHostClash GETs the node's route array and reports whether
-// route_<routeID> is already present, and whether any OTHER route shares a host
-// with `hosts` (in which case a POST-append could mis-order path-vs-root match
-// and we must fall back to a full /load to preserve deterministic id-order).
-func (s *Service) routePresenceAndHostClash(ctx context.Context, client *caddyapi.Client, routeID int64, hosts []string) (present, sharesHost bool, err error) {
-	raw, err := client.GetRaw(ctx, "/config/apps/http/servers/srv0/routes")
-	if err != nil {
-		return false, false, err
-	}
-	trimmed := strings.TrimSpace(string(raw))
-	if trimmed == "" || trimmed == "null" {
-		return false, false, nil // no routes on the node yet
-	}
-	var arr []map[string]any
-	if err := json.Unmarshal(raw, &arr); err != nil {
-		return false, false, err
-	}
-	caddyID := fmt.Sprintf("route_%d", routeID)
-	for _, obj := range arr {
-		if id, _ := obj["@id"].(string); id == caddyID {
-			present = true
-			continue
-		}
-		// Wildcard-aware: an existing "*.example.com" catch-all also matches a
-		// new "app.example.com" and would shadow it if we merely appended.
-		if caddyapi.HostSetsOverlap(hosts, routeMatchHosts(obj)) {
-			sharesHost = true
-		}
-	}
-	return present, sharesHost, nil
-}
-
-// pushRouteIncremental applies a single-route change to one node via Caddy @id
-// endpoints, avoiding a whole-config /load. ANY failure (probe, build, HTTP, or
-// an unsafe-ordering condition) falls back to a full pushNodeConfigLocked so
-// behavior is never worse than a /load. last_pushed_hash is intentionally not
-// rewritten here (it is write-only/unused; drift rebuilds from DB and the
-// incremental object is byte-identical to a /load element, so drift is unaffected).
-func (s *Service) pushRouteIncremental(ctx context.Context, nodeID, routeID int64, op routeOp) error {
-	if !s.IncrementalPush {
-		return s.pushNodeConfig(ctx, nodeID)
-	}
-	// Wildcard routes drive tls.automation.policies, which lives outside the
-	// per-route @id surface; an incremental op would never emit the DNS-01
-	// policy. Force a full /load so the policy set re-derives. Cheap (rare).
-	if s.routeIsWildcard(ctx, routeID) {
-		return s.pushNodeConfig(ctx, nodeID)
-	}
-	var apiURL string
-	if err := s.DB.QueryRowContext(ctx, "SELECT api_url FROM caddy_nodes WHERE id = ?", nodeID).Scan(&apiURL); err != nil {
-		return err
-	}
-	client := caddyapi.New(apiURL)
-	caddyID := fmt.Sprintf("route_%d", routeID)
-
-	lock := s.nodeLock(nodeID)
-	lock.Lock()
-	defer lock.Unlock()
-
-	switch op {
-	case routeRemove:
-		if err := client.DeleteRoute(ctx, caddyID); err != nil {
-			if isNotFound(err) {
-				return nil // already gone == desired state
-			}
-			s.Logger.Warn("incremental delete failed, full resync", "node_id", nodeID, "route_id", routeID, "err", err)
-			return s.pushNodeConfigLocked(ctx, nodeID)
-		}
-		return nil
-
-	case routeUpsert:
-		built, ok, berr := s.buildOneRoute(ctx, nodeID, routeID)
-		if berr != nil {
-			return s.pushNodeConfigLocked(ctx, nodeID)
-		}
-		if !ok {
-			// Not eligible: ensure it is absent on the node, then done.
-			if derr := client.DeleteRoute(ctx, caddyID); derr != nil && !isNotFound(derr) {
-				return s.pushNodeConfigLocked(ctx, nodeID)
-			}
-			return nil
-		}
-		obj := caddyapi.BuildRoute(built)
-		present, sharesHost, perr := s.routePresenceAndHostClash(ctx, client, routeID, built.Hosts)
-		switch {
-		case perr != nil:
-			return s.pushNodeConfigLocked(ctx, nodeID)
-		case present:
-			// Replace in place (preserves index/order) - shape-agnostic.
-			if err := client.ReplaceRoute(ctx, caddyID, obj); err != nil {
-				s.Logger.Warn("incremental replace failed, full resync", "node_id", nodeID, "route_id", routeID, "err", err)
-				return s.pushNodeConfigLocked(ctx, nodeID)
-			}
-		case sharesHost:
-			return s.pushNodeConfigLocked(ctx, nodeID) // keep deterministic order
-		default:
-			if err := client.AddRoute(ctx, obj); err != nil {
-				s.Logger.Warn("incremental add failed, full resync", "node_id", nodeID, "route_id", routeID, "err", err)
-				return s.pushNodeConfigLocked(ctx, nodeID)
-			}
-		}
-		if s.Metrics != nil {
-			s.Metrics.CaddyPushOK()
-		}
-		return nil
-	}
-	return nil
-}
-
-// loadErrorBranding pulls the per-install branding bits used by
-// Caddy-served error / maintenance pages. Empty struct on any failure
-// so the renderer falls back to neutral defaults rather than panicking
-// the resync.
-func (s *Service) loadErrorBranding(ctx context.Context) caddyapi.ErrorBranding {
-	b := caddyapi.ErrorBranding{}
-	if s.DB == nil {
-		return b
-	}
-	c, cancel := context.WithTimeout(ctx, 2*time.Second)
-	defer cancel()
-	rows, err := s.DB.QueryContext(c,
-		"SELECT `key`, value FROM settings WHERE `key` IN ("+
-			"'branding.brand_name',"+
-			"'branding.error_logo_url','branding.error_logo_link','branding.error_bg_color')")
-	if err != nil {
-		return b
-	}
-	defer rows.Close()
-	for rows.Next() {
-		var k, v string
-		if err := rows.Scan(&k, &v); err != nil {
-			continue
-		}
-		switch k {
-		case "branding.brand_name":
-			b.Brand = v
-		case "branding.error_logo_url":
-			b.LogoURL = v
-		case "branding.error_logo_link":
-			b.LogoLink = v
-		case "branding.error_bg_color":
-			b.BgColor = v
-		}
-	}
-	return b
-}
-
-// geoBlockCfg is the resolved geo-block response config for a route: either the
-// owning client's customisation or the panel-wide default.
-type geoBlockCfg struct {
-	action, redirectURL, title, message, logoURL, bgColor string
-}
-
-// loadGeoBlockDefault reads the panel-wide geo-block response default from the
-// settings KV table. Used as the fallback when a client has not customised it.
-func (s *Service) loadGeoBlockDefault(ctx context.Context) geoBlockCfg {
-	var g geoBlockCfg
-	if s.DB == nil {
-		return g
-	}
-	c, cancel := context.WithTimeout(ctx, 2*time.Second)
-	defer cancel()
-	rows, err := s.DB.QueryContext(c,
-		"SELECT `key`, value FROM settings WHERE `key` IN ("+
-			"'geoblock.action','geoblock.redirect_url','geoblock.title',"+
-			"'geoblock.message','geoblock.logo_url','geoblock.bg_color')")
-	if err != nil {
-		return g
-	}
-	defer rows.Close()
-	for rows.Next() {
-		var k, v string
-		if err := rows.Scan(&k, &v); err != nil {
-			continue
-		}
-		switch k {
-		case "geoblock.action":
-			g.action = v
-		case "geoblock.redirect_url":
-			g.redirectURL = v
-		case "geoblock.title":
-			g.title = v
-		case "geoblock.message":
-			g.message = v
-		case "geoblock.logo_url":
-			g.logoURL = v
-		case "geoblock.bg_color":
-			g.bgColor = v
-		}
-	}
-	return g
-}
-
-// loadTrustCloudflareIP reads the global cloudflare.trust_connecting_ip setting
-// (default false). Same key the panel's own CloudflareIP middleware honours, so
-// one toggle covers both the panel and the nodes.
-func (s *Service) loadTrustCloudflareIP(ctx context.Context) bool {
-	if s.DB == nil {
-		return false
-	}
-	c, cancel := context.WithTimeout(ctx, 2*time.Second)
-	defer cancel()
-	var v string
-	_ = s.DB.QueryRowContext(c, "SELECT value FROM settings WHERE `key` = 'cloudflare.trust_connecting_ip'").Scan(&v)
-	return v == "1"
-}
-
-// loadMTLSFailOpen reads the global mtls.fail_open setting (default false = fail closed).
-func (s *Service) loadMTLSFailOpen(ctx context.Context) bool {
-	if s.DB == nil {
-		return false
-	}
-	c, cancel := context.WithTimeout(ctx, 2*time.Second)
-	defer cancel()
-	var v string
-	_ = s.DB.QueryRowContext(c, "SELECT value FROM settings WHERE `key` = 'mtls.fail_open'").Scan(&v)
-	return v == "1"
-}
-
-// resolveRandomEgressIP picks a stable IP for a route from the node's inventory.
-// Uses route ID mod len(ips) so the same route always maps to the same IP across rebuilds.
-func resolveRandomEgressIP(ips []string, routeID int64) string {
-	if len(ips) == 0 {
-		return ""
-	}
-	return ips[int(routeID)%len(ips)]
-}
-
-// buildRoutesForNode collects every active/dns_ok/pending_ssl route placed on
-// the given node, applies plan overrides, and returns Caddy route structs.
-func (s *Service) buildRoutesForNode(ctx context.Context, nodeID int64) ([]caddyapi.Route, []int64, error) {
-	// Fetch node's outbound IP inventory for 'random' egress mode resolution.
-	var nodeOutboundIPsJSON sql.NullString
-	_ = s.DB.QueryRowContext(ctx,
-		`SELECT outbound_ips FROM caddy_nodes WHERE id = ?`, nodeID,
-	).Scan(&nodeOutboundIPsJSON)
-	var nodeOutboundIPs []string
-	if nodeOutboundIPsJSON.Valid && nodeOutboundIPsJSON.String != "" {
-		_ = json.Unmarshal([]byte(nodeOutboundIPsJSON.String), &nodeOutboundIPs)
-	}
-	// Panel-wide geo-block default + brand name, used when a client has not
-	// customised its own geo-block response. Loaded once per build.
-	geoDefault := s.loadGeoBlockDefault(ctx)
-	geoBrand := s.loadErrorBranding(ctx).Brand
-	// Operator's fail-open choice governs whether a route whose mTLS enforcement
-	// cannot be emitted is served open or denied. Loaded once per build.
-	mtlsFailOpen := s.loadMTLSFailOpen(ctx)
-	rows, err := s.DB.QueryContext(ctx,
-		`SELECT r.id, r.domain, COALESCE(r.aliases,''), COALESCE(r.aliases_verified,''), r.path_prefix, r.upstream_port, r.upstream_scheme, r.upstream_skip_tls_verify,
-		        r.websocket, r.force_https,
-		        r.http2_enabled, r.http3_enabled, r.ssl_enabled,
-		        -- Per-route backend_ip_override beats both peer IP and the
-		        -- shared service backend_ip, so editing one route doesn't
-		        -- change every sibling route that JOINs the same service.
-		        COALESCE(NULLIF(r.backend_ip_override, ''), p_use.assigned_ip, sv.backend_ip),
-		        COALESCE(p_use.assigned_ip, ''),
-		        r.kind, COALESCE(r.redirect_url,''), COALESCE(r.redirect_code,0),
-		        r.cache_enabled, r.cache_ttl_secs, COALESCE(r.cache_public,0), COALESCE(r.custom_headers,''),
-		        r.maintenance_mode, COALESCE(r.maintenance_message,''),
-		        COALESCE(r.cache_vary,''),
-		        COALESCE(r.access_allow,''), COALESCE(r.access_deny,''),
-		        COALESCE(r.access_block_all, 0), COALESCE(r.maintenance_allow,''),
-		        COALESCE(r.custom_config,''),
-		        r.via_wg_peer_id, p_use.status,
-		        COALESCE(r.basic_auth_user,''), COALESCE(r.basic_auth_bcrypt,''),
-		        COALESCE(r.sso_provider_url,''), COALESCE(r.sso_copy_headers,''), COALESCE(r.sso_trusted_proxies,''),
-		        COALESCE(r.sso_paths,''), COALESCE(r.sso_hosts,''),
-		        COALESCE(r.sso_strict_mode,0),
-		        COALESCE(sso_peer.assigned_ip, ''),
-	        -- Built-in portal: gate only when the flag is on AND >=1 grant exists,
-	        -- otherwise the route would be gated with nobody allowed (effective
-	        -- lockout). The verifier still re-checks membership per request.
-	        (COALESCE(r.portal_protect,0)=1 AND EXISTS(SELECT 1 FROM route_access_grants rag WHERE rag.route_id=r.id)),
-	        COALESCE(r.upstream_external, 0), COALESCE(r.upstream_host_header, ''), COALESCE(r.proxy_secret_enc, ''),
-	        COALESCE(r.compress_disabled, 0),
-	        COALESCE(r.lb_policy,''),
-	        COALESCE(r.lb_header_field,''), COALESCE(r.lb_cookie_name,''), COALESCE(r.lb_cookie_secret,''),
-	        COALESCE(r.health_active_uri,''), COALESCE(r.health_active_interval,10), COALESCE(r.health_active_timeout,5),
-	        COALESCE(r.health_active_status,0), COALESCE(r.health_active_fails,3),
-	        COALESCE(r.health_passive_enabled,0), COALESCE(r.health_passive_fail_dur,30), COALESCE(r.health_passive_max_fail,3),
-	        COALESCE(r.lb_try_duration_ms,5000), COALESCE(r.lb_try_interval_ms,250),
-	        COALESCE(r.rate_enabled,0), COALESCE(r.rate_window,''), COALESCE(r.rate_max_events,0), COALESCE(r.rate_key,''),
-	        COALESCE(r.waf_enabled,0), COALESCE(r.waf_blocking,0), COALESCE(r.waf_directives,''),
-	        COALESCE(r.geo_mode,'off'), COALESCE(r.geo_countries,''),
-	        COALESCE(r.geo_response_code,403), COALESCE(r.geo_fail_closed,0), COALESCE(r.geo_allow_cidrs,''),
-	        COALESCE(r.geo_continents,''), r.geo_block_cidrs,
-	        COALESCE(r.error_override,0), COALESCE(r.error_html,''), COALESCE(r.error_logo_url,''),
-	        COALESCE(r.error_brand,''), COALESCE(r.error_bg_color,''),
-	        COALESCE(r.outbound_ip_mode,'default'), COALESCE(r.outbound_ip,''),
-	        COALESCE(pl.allow_egress_ip, 0),
-	        COALESCE(r.dns_resolver_ip,''), COALESCE(dns_peer.assigned_ip,''),
-	        COALESCE(r.dns_address_family,'any'),
-	        COALESCE(r.require_client_cert,0), COALESCE(mca.cert_pem,''),
-	        COALESCE(r.dial_timeout_ms,0), COALESCE(r.response_header_timeout_ms,0),
-	        COALESCE(cl.geo_block_action,''), COALESCE(cl.geo_block_redirect_url,''),
-	        COALESCE(cl.geo_block_title,''), COALESCE(cl.geo_block_message,''),
-	        COALESCE(cl.geo_block_logo_url,''), COALESCE(cl.geo_block_bg_color,'')
-		 FROM routes r
-		 JOIN services sv ON sv.id = r.service_id
-		 LEFT JOIN clients cl ON cl.id = sv.client_id
-		 LEFT JOIN plans pl ON pl.id = sv.plan_id
-		 LEFT JOIN mtls_cas mca ON mca.id = r.mtls_ca_id AND mca.status = 'active'
-		 LEFT JOIN customer_wg_peer p_base
-		   ON p_base.id = r.via_wg_peer_id
-		 LEFT JOIN customer_wg_peer p_use ON (
-		     (p_base.peer_group_id IS NOT NULL
-		         AND p_use.peer_group_id = p_base.peer_group_id
-		         AND p_use.node_id = ?
-		         AND p_use.status <> 'revoked')
-		     OR (p_base.peer_group_id IS NULL
-		         AND p_use.id = r.via_wg_peer_id
-		         AND p_use.status <> 'revoked')
-		 )
-		 LEFT JOIN customer_wg_peer sso_peer
-		   ON sso_peer.id = r.sso_via_wg_peer_id
-		      AND sso_peer.status <> 'revoked'
-		 -- DNS peer is node-scoped like p_use: fan-out nodes must query their
-		 -- own group member, not the primary's peer IP.
-		 LEFT JOIN customer_wg_peer dns_base
-		   ON dns_base.id = r.dns_resolver_via_wg_peer_id
-		 LEFT JOIN customer_wg_peer dns_peer ON (
-		     (dns_base.peer_group_id IS NOT NULL
-		         AND dns_peer.peer_group_id = dns_base.peer_group_id
-		         AND dns_peer.node_id = ?
-		         AND dns_peer.status <> 'revoked')
-		     OR (dns_base.peer_group_id IS NULL
-		         AND dns_peer.id = r.dns_resolver_via_wg_peer_id
-		         AND dns_peer.status <> 'revoked')
-		 )
-		 -- Anchor routes plus fan-out copies (active_active/failover peers live
-		 -- only in route_node_assignments; without this peers pushed routes: 0).
-		 WHERE (r.caddy_node_id = ?
-		        OR EXISTS (SELECT 1 FROM route_node_assignments rna
-		                    WHERE rna.route_id = r.id AND rna.node_id = ?))
-		   AND r.status IN ('dns_ok','active','pending_ssl')
-		   -- Defence in depth: only advanceRoute should be able to put a route
-		   -- in a serving status, and it refuses unverified ones. A status set
-		   -- by any other path must still not emit an unproven host matcher.
-		   AND COALESCE(r.domain_verified, 0) = 1
-		 ORDER BY r.id ASC`, nodeID, nodeID, nodeID, nodeID)
-	if err != nil {
-		return nil, nil, err
-	}
-	defer rows.Close()
-
-	var built []caddyapi.Route
-	var ids []int64
-	for rows.Next() {
-		var (
-			id                             int64
-			domain, aliases                string
-			aliasesVerified                string
-			path                           string
-			port                           int
-			scheme                         string
-			skipTLS                        bool
-			ws, fhttps, h2, h3, sslEnabled bool
-			ip                             string
-			tunnelResolverIP               string
-			kind                           string
-			redirURL                       string
-			redirCode                      int
-			cacheEnabled                   bool
-			cachePublic                    bool
-			cacheTTL                       int
-			headersJSON                    string
-			maintMode                      bool
-			maintMsg                       string
-			cacheVary                      string
-			accessAllow                    string
-			accessDeny                     string
-			accessBlockAll                 bool
-			maintenanceAllow               string
-			customCfg                      string
-		)
-		var viaPeerID sql.NullInt64
-		var peerStatus sql.NullString
-		var baUser, baHash string
-		var ssoProviderURL, ssoCopyHeadersRaw, ssoTrustedProxies string
-		var ssoPathsRaw, ssoHostsRaw string
-		var ssoStrictMode bool
-		var ssoResolverIP string
-		var portalProtect bool
-		var upstreamExternal bool
-		var upstreamHostHeader, proxySecretEnc string
-		var compressDisabled bool
-		var lbPolicy string
-		var lbHeaderField, lbCookieName, lbCookieSecret string
-		var hActiveURI string
-		var hActiveInterval, hActiveTimeout, hActiveStatus, hActiveFails int
-		var hPassiveEnabled bool
-		var hPassiveFailDur, hPassiveMaxFail int
-		var lbTryDurationMs, lbTryIntervalMs int
-		var rateEnabled bool
-		var rateWindow, rateKey string
-		var rateMaxEvents int
-		var wafEnabled, wafBlocking bool
-		var wafDirectives string
-		var geoMode, geoCountries string
-		var geoResponseCode int
-		var geoFailClosed bool
-		var geoAllowCIDRs string
-		var geoContinents string
-		var geoBlockCIDRs sql.NullString
-		var errOverride bool
-		var errHTML, errLogo, errBrand, errBg string
-		var outboundIPMode, outboundIP string
-		var planAllowEgress bool
-		var dnsResolverIP, dnsResolverPeerIP, dnsAddressFamily string
-		var requireClientCert bool
-		var mtlsCACertPEM string
-		var dialTimeoutMs, responseHeaderTimeoutMs int
-		var clGeoAction, clGeoRedirect, clGeoTitle, clGeoMessage, clGeoLogo, clGeoBg string
-		if err := rows.Scan(&id, &domain, &aliases, &aliasesVerified, &path, &port, &scheme, &skipTLS, &ws, &fhttps, &h2, &h3, &sslEnabled, &ip,
-			&tunnelResolverIP,
-			&kind, &redirURL, &redirCode, &cacheEnabled, &cacheTTL, &cachePublic, &headersJSON,
-			&maintMode, &maintMsg, &cacheVary, &accessAllow, &accessDeny,
-			&accessBlockAll, &maintenanceAllow, &customCfg,
-			&viaPeerID, &peerStatus, &baUser, &baHash,
-			&ssoProviderURL, &ssoCopyHeadersRaw, &ssoTrustedProxies,
-			&ssoPathsRaw, &ssoHostsRaw,
-			&ssoStrictMode,
-			&ssoResolverIP,
-			&portalProtect,
-			&upstreamExternal, &upstreamHostHeader, &proxySecretEnc,
-			&compressDisabled,
-			&lbPolicy,
-			&lbHeaderField, &lbCookieName, &lbCookieSecret,
-			&hActiveURI, &hActiveInterval, &hActiveTimeout, &hActiveStatus, &hActiveFails,
-			&hPassiveEnabled, &hPassiveFailDur, &hPassiveMaxFail,
-			&lbTryDurationMs, &lbTryIntervalMs,
-			&rateEnabled, &rateWindow, &rateMaxEvents, &rateKey,
-			&wafEnabled, &wafBlocking, &wafDirectives,
-			&geoMode, &geoCountries,
-			&geoResponseCode, &geoFailClosed, &geoAllowCIDRs,
-			&geoContinents, &geoBlockCIDRs,
-			&errOverride, &errHTML, &errLogo, &errBrand, &errBg,
-			&outboundIPMode, &outboundIP, &planAllowEgress,
-			&dnsResolverIP, &dnsResolverPeerIP, &dnsAddressFamily,
-			&requireClientCert, &mtlsCACertPEM,
-			&dialTimeoutMs, &responseHeaderTimeoutMs,
-			&clGeoAction, &clGeoRedirect, &clGeoTitle, &clGeoMessage, &clGeoLogo, &clGeoBg); err != nil {
-			return nil, nil, err
-		}
-		// Re-check plan entitlement at build time so revoking the flag takes effect immediately.
-		if !planAllowEgress {
-			outboundIPMode = "default"
-			outboundIP = ""
-		}
-		var ssoCopyHeaders []string
-		for _, h := range strings.FieldsFunc(ssoCopyHeadersRaw, func(r rune) bool { return r == '\n' || r == '\r' || r == ',' }) {
-			if v := strings.TrimSpace(h); v != "" {
-				ssoCopyHeaders = append(ssoCopyHeaders, v)
-			}
-		}
-		var ssoTrusted []string
-		for _, p := range strings.FieldsFunc(ssoTrustedProxies, func(r rune) bool { return r == ',' || r == ' ' || r == '\t' }) {
-			if v := strings.TrimSpace(p); v != "" {
-				ssoTrusted = append(ssoTrusted, v)
-			}
-		}
-		// Resolve 'random' egress to a stable concrete IP so static Caddy config can use it.
-		if outboundIPMode == "random" {
-			if picked := resolveRandomEgressIP(nodeOutboundIPs, id); picked != "" {
-				outboundIP = picked
-			} else {
-				outboundIPMode = "default"
-			}
-		}
-		// Skip routes pointing at a missing or revoked tunnel rather than
-		// silently falling back to the static backend_ip - Caddy returning
-		// 502 is preferable to "huh, my traffic suddenly bypassed the VPN".
-		if viaPeerID.Valid && (!peerStatus.Valid || peerStatus.String == "revoked") {
-			s.Logger.Warn("skipping route with revoked/missing tunnel",
-				"route_id", id, "domain", domain, "peer_id", viaPeerID.Int64)
-			continue
-		}
-		// Only PROVEN aliases become host matchers: an alias whose ownership was
-		// never demonstrated would otherwise intercept the real owner's traffic.
-		proven := map[string]bool{}
-		for _, a := range splitHostList(aliasesVerified) {
-			proven[a] = true
-		}
-		hosts := []string{domain}
-		for _, a := range splitHostList(aliases) {
-			if a != domain && proven[a] {
-				hosts = append(hosts, a)
-			}
-		}
-		var vary []string
-		if cacheVary != "" {
-			for _, p := range strings.Split(cacheVary, ",") {
-				if v := strings.TrimSpace(p); v != "" {
-					vary = append(vary, v)
-				}
-			}
-		}
-		splitCIDRs := func(s string) []string {
-			if s == "" {
-				return nil
-			}
-			out := []string{}
-			for _, p := range strings.FieldsFunc(s, func(r rune) bool {
-				return r == ',' || r == ' ' || r == '\t' || r == '\n' || r == '\r'
-			}) {
-				if v := strings.TrimSpace(p); v != "" {
-					out = append(out, v)
-				}
-			}
-			return out
-		}
-		allowList := splitCIDRs(accessAllow)
-		denyList := splitCIDRs(accessDeny)
-		maintAllowList := splitCIDRs(maintenanceAllow)
-		if !sslEnabled {
-			fhttps = false
-		}
-		var headers map[string]string
-		if headersJSON != "" {
-			_ = json.Unmarshal([]byte(headersJSON), &headers)
-		}
-		// Hostname-via-tunnel-DNS feature is disabled at build time.
-		// External route: the upstream FQDN (ip) is intentionally a hostname,
-		// re-enforce the allowlist and decrypt the inbound bearer. Skip the
-		// route entirely (never emit an ungated/disallowed external proxy)
-		// if either fails - an emitted route without its gate is an open relay.
-		external := upstreamExternal
-		proxySecret := ""
-		if external {
-			if !s.externalHostAllowed(ip) {
-				s.Logger.Warn("external route host not allowlisted, skipping", "route_id", id, "host", ip)
-				continue
-			}
-			if proxySecretEnc != "" {
-				if s.DecryptSecret == nil {
-					s.Logger.Error("external route secret undecryptable (no key), skipping", "route_id", id)
-					continue
-				}
-				sec, derr := s.DecryptSecret(proxySecretEnc)
-				if derr != nil {
-					s.Logger.Error("external route secret decrypt failed, skipping", "route_id", id, "err", derr)
-					continue
-				}
-				proxySecret = sec
-			}
-		}
-		// Hostname backend over tunnel needs a DNS resolver (dynamic
-		// upstreams resolve it via the peer). Without one, fall back to
-		// peer IP so the route doesn't 502 on an unresolvable name.
-		// External routes exempt - their upstream is a public FQDN.
-		if !external && tunnelResolverIP != "" && ip != "" && !looksLikeIP(ip) &&
-			dnsResolverIP == "" && dnsResolverPeerIP == "" {
-			ip = tunnelResolverIP
-		}
-		backendResolver := ""
-		// lb_cookie_secret is stored encrypted at rest (SECRET-02); decrypt for
-		// the Caddy push. Legacy plaintext rows fall through unchanged.
-		if lbCookieSecret != "" && s.DecryptSecret != nil {
-			if dec, derr := s.DecryptSecret(lbCookieSecret); derr == nil {
-				lbCookieSecret = dec
-			}
-		}
-		// Resolve geo-block response: a client's own customisation wins over the
-		// panel-wide default (empty client action = inherit the default).
-		gb := geoDefault
-		if strings.TrimSpace(clGeoAction) != "" {
-			gb = geoBlockCfg{clGeoAction, clGeoRedirect, clGeoTitle, clGeoMessage, clGeoLogo, clGeoBg}
-		}
-		// An enabled auth gate that cannot be emitted must not silently vanish.
-		// mTLS respects the operator's mtls.fail_open; the portal never does.
-		portalReady := s.PanelInternalHost != "" && s.PanelInternalPort != 0
-		mtlsEnforceable := sslEnabled && mtlsCACertPEM != "" && caddyapi.MTLSCAUsable(mtlsCACertPEM)
-		built = append(built, caddyapi.Route{
-			ID:                    fmt.Sprintf("%d", id),
-			Hosts:                 hosts,
-			PathPrefix:            path,
-			UpstreamIP:            ip,
-			UpstreamPort:          port,
-			BackendResolver:       backendResolver,
-			UpstreamScheme:        scheme,
-			UpstreamSkipTLSVerify: skipTLS,
-			WebSocket:             ws,
-			ForceHTTPS:            fhttps,
-			HTTP2:                 h2,
-			HTTP3:                 h3,
-			Headers:               headers,
-			Kind:                  kind,
-			RedirectURL:           redirURL,
-			RedirectCode:          redirCode,
-			CacheEnabled:          cacheEnabled,
-			CachePublic:           cachePublic,
-			CacheTTLSeconds:       cacheTTL,
-			CacheVary:             vary,
-			MaintenanceMode:       maintMode,
-			MaintenanceMessage:    maintMsg,
-			AccessAllow:           allowList,
-			AccessDeny:            denyList,
-			AccessBlockAll:        accessBlockAll,
-			MaintenanceAllow:      maintAllowList,
-			CustomHandlers:        customCfg,
-			BasicAuthUser:         baUser,
-			BasicAuthBcrypt:       baHash,
-			SSOProviderURL:        ssoProviderURL,
-			SSOCopyHeaders:        ssoCopyHeaders,
-			SSOTrustedProxies:     ssoTrusted,
-			SSOPaths:              splitCIDRs(ssoPathsRaw),
-			SSOHosts:              splitCIDRs(ssoHostsRaw),
-			// SSO-via-tunnel: peer IP is used as static dial host (port
-			// comes from SSO Provider URL). No DNS lookup attempted - peer
-			// must expose the IdP port on its host network.
-			SSOResolver:   ssoResolverIP,
-			SSOStrictMode: ssoStrictMode,
-
-			// Built-in forward-auth portal. Verify + login are dialed at the
-			// panel (same internal host:port as the self-bootstrap route). Plain
-			// HTTP over the internal network; the panel sets the protected-host
-			// cookie itself. Skipped entirely when PanelInternalHost is unset.
-			PortalProtect: portalProtect && portalReady,
-			PortalDial:    s.portalDial(),
-			// Portal is an explicit per-route opt-in: no verifier means deny,
-			// never silently serve the protected backend to the public.
-			PortalDenyOnMisconfig: portalProtect && !portalReady,
-
-			// External HTTPS upstream: SNI + Host both use the stored header
-			// (falls back to the FQDN in the builder); ProxySecret gates inbound.
-			External:                external,
-			UpstreamSNI:             upstreamHostHeader,
-			UpstreamHostHeader:      upstreamHostHeader,
-			ProxySecret:             proxySecret,
-			CompressDisabled:        compressDisabled,
-			LBPolicy:                lbPolicy,
-			LBHeaderField:           lbHeaderField,
-			LBCookieName:            lbCookieName,
-			LBCookieSecret:          lbCookieSecret,
-			WeightedLBAvailable:     s.WeightedLBAvailable,
-			LBTryDurationMs:         lbTryDurationMs,
-			LBTryIntervalMs:         lbTryIntervalMs,
-			DialTimeoutMs:           dialTimeoutMs,
-			ResponseHeaderTimeoutMs: responseHeaderTimeoutMs,
-			HealthURI:               hActiveURI,
-			HealthIntervalSecs:      hActiveInterval,
-			HealthTimeoutSecs:       hActiveTimeout,
-			HealthExpectStatus:      hActiveStatus,
-			HealthFails:             hActiveFails,
-			HealthPassive:           hPassiveEnabled,
-			HealthFailDurationSecs:  hPassiveFailDur,
-			HealthMaxFails:          hPassiveMaxFail,
-
-			RateLimitEnabled:         rateEnabled,
-			RateLimitWindow:          rateWindow,
-			RateLimitMaxEvents:       rateMaxEvents,
-			RateLimitKey:             rateKey,
-			RateLimitModuleAvailable: s.RateLimitModuleAvailable,
-			WAFEnabled:               wafEnabled,
-			WAFBlocking:              wafBlocking,
-			WAFDirectives:            wafDirectives,
-			WAFModuleAvailable:       s.WAFModuleAvailable,
-			GeoMode:                  geoMode,
-			GeoCountries:             geoCountries,
-			GeoModuleAvailable:       s.GeoModuleAvailable && geoip.HasCountryDB(),
-			GeoResponseCode:          geoResponseCode,
-			GeoFailClosed:            geoFailClosed,
-			GeoAllowCIDRs:            geoAllowCIDRs,
-			GeoContinents:            geoContinents,
-			GeoBlockCIDRs:            geoBlockCIDRs.String,
-			GeoBlockAction:           gb.action,
-			GeoRedirectURL:           gb.redirectURL,
-			GeoBlockTitle:            gb.title,
-			GeoBlockMessage:          gb.message,
-			GeoBlockBranding:         caddyapi.ErrorBranding{Brand: geoBrand, LogoURL: gb.logoURL, BgColor: gb.bgColor},
-
-			// Per-route error/maintenance page override (else node-wide branding).
-			CustomErrorOverride: errOverride,
-			CustomErrorHTML:     errHTML,
-			CustomErrorBranding: caddyapi.ErrorBranding{LogoURL: errLogo, Brand: errBrand, BgColor: errBg},
-
-			OutboundIPMode: outboundIPMode,
-			OutboundIP:     outboundIP,
-
-			DNSResolverIP:          dnsResolverIP,
-			DNSResolverViaWGPeerIP: dnsResolverPeerIP,
-			DNSAddressFamily:       dnsAddressFamily,
-
-			// mTLS client-cert enforcement only over TLS, and only when the
-			// selected CA is still active AND parsable (JOIN yields '' otherwise).
-			RequireClientCert: requireClientCert && mtlsEnforceable,
-			MTLSCACertPEM:     mtlsCACertPEM,
-			// No enforceable policy + fail-closed operator setting = deny the
-			// route instead of serving it with no client-cert requirement.
-			MTLSDenyOnMisconfig: requireClientCert && !mtlsEnforceable && !mtlsFailOpen,
-			PanelBaseURL:        panelBaseURL(s.AskURL),
-		})
-		// Audit the quarantine: BuildRoute replaces the whole route with a 503,
-		// so without this the operator sees an outage with no stated cause.
-		if reason := caddyapi.CustomHandlerQuarantine(built[len(built)-1]); reason != "" {
-			s.Logger.Error("route quarantined: custom handler chain rejected",
-				"route_id", id, "domain", domain, "reason", reason)
-			audit.Write(ctx, s.DB, s.Logger, nil, audit.Entry{
-				ActorType: audit.ActorSystem,
-				Action:    "route.custom_handlers.quarantined",
-				Entity:    "route",
-				EntityID:  fmt.Sprintf("%d", id),
-				Meta:      map[string]any{"domain": domain, "reason": reason},
-			})
-		}
-		// Audit when require_client_cert=1 but enforcement is skipped (no active
-		// CA, SSL off, unparsable PEM, or mtls_ca_id NULL).
-		if requireClientCert && !mtlsEnforceable {
-			audit.Write(ctx, s.DB, s.Logger, nil, audit.Entry{
-				ActorType: audit.ActorSystem,
-				Action:    "route.mtls.pushed_without_tls",
-				Entity:    "route",
-				EntityID:  fmt.Sprintf("%d", id),
-				Meta:      map[string]any{"domain": domain, "fail_open": mtlsFailOpen},
-			})
-		}
-		ids = append(ids, id)
-	}
-	// Attach additional backends (route_upstreams) in one batched query to
-	// avoid N+1; zero-row routes keep their single-dial primary. Best-effort:
-	// a query error leaves routes single-dial rather than failing the build.
-	s.attachRouteUpstreams(ctx, built, ids)
-	s.attachLocationRules(ctx, built, ids)
-	s.attachBasicAuthUsers(ctx, built, ids)
-	s.attachMTLSPathRules(ctx, built, ids)
-	s.attachRBACTokens(built, ids, nodeID)
-	// Emission order, not DB id order: a catch-all ahead of a narrower sibling
-	// on the same host would shadow it. ids must follow the same permutation -
-	// buildOneRoute pairs ids[i] with built[i].
-	if ord := caddyapi.EmissionOrder(built); len(ids) == len(built) {
-		sortedBuilt := make([]caddyapi.Route, len(built))
-		sortedIDs := make([]int64, len(ids))
-		for k, i := range ord {
-			sortedBuilt[k] = built[i]
-			sortedIDs[k] = ids[i]
-		}
-		built, ids = sortedBuilt, sortedIDs
-	}
-	return built, ids, nil
-}
-
-// attachBasicAuthUsers loads route_basic_auth_users in one IN(...) query and
-// maps them onto the built routes. Routes with users get BasicAuthUsers set;
-// routes without rows keep the legacy single-user BasicAuthUser/BasicAuthBcrypt.
-func (s *Service) attachBasicAuthUsers(ctx context.Context, built []caddyapi.Route, ids []int64) {
-	if len(ids) == 0 {
-		return
-	}
-	idx := make(map[int64]int, len(ids))
-	ph := make([]string, len(ids))
-	args := make([]any, len(ids))
-	for i, id := range ids {
-		idx[id] = i
-		ph[i] = "?"
-		args[i] = id
-	}
-	rows, err := s.DB.QueryContext(ctx,
-		`SELECT route_id, username, bcrypt_hash
-		   FROM route_basic_auth_users
-		  WHERE route_id IN (`+strings.Join(ph, ",")+`)
-		  ORDER BY route_id, username ASC`, args...)
-	if err != nil {
-		if s.Logger != nil {
-			s.Logger.Warn("route_basic_auth_users load failed; using single-user fallback", "err", err)
-		}
-		return
-	}
-	defer rows.Close()
-	for rows.Next() {
-		var rid int64
-		var u caddyapi.BasicAuthUser
-		if err := rows.Scan(&rid, &u.Username, &u.Hash); err != nil {
-			continue
-		}
-		if i, ok := idx[rid]; ok {
-			built[i].BasicAuthUsers = append(built[i].BasicAuthUsers, u)
-		}
-	}
-}
-
-// attachMTLSPathRules loads mtls_path_rules for the given route IDs in one
-// batch query and sets MTLSPathRules on each matching route.
-func (s *Service) attachMTLSPathRules(ctx context.Context, built []caddyapi.Route, ids []int64) {
-	if len(ids) == 0 {
-		return
-	}
-	idx := make(map[int64]int, len(ids))
-	ph := make([]string, len(ids))
-	args := make([]any, len(ids))
-	for i, id := range ids {
-		idx[id] = i
-		ph[i] = "?"
-		args[i] = id
-	}
-	rows, err := s.DB.QueryContext(ctx,
-		`SELECT pr.route_id, ro.name, pr.path_pattern
-		   FROM mtls_path_rules pr
-		   JOIN mtls_roles ro ON ro.id = pr.required_role_id
-		  WHERE pr.route_id IN (`+strings.Join(ph, ",")+`)
-		  ORDER BY pr.route_id, pr.id ASC`, args...)
-	if err != nil {
-		if s.Logger != nil {
-			s.Logger.Warn("mtls_path_rules load failed", "err", err)
-		}
-		return
-	}
-	defer rows.Close()
-	for rows.Next() {
-		var rid int64
-		var role, pattern string
-		if err := rows.Scan(&rid, &role, &pattern); err != nil {
-			continue
-		}
-		if i, ok := idx[rid]; ok {
-			built[i].MTLSPathRules = append(built[i].MTLSPathRules, caddyapi.MTLSPathRule{
-				PathPattern:  pattern,
-				RequiredRole: role,
-			})
-		}
-	}
-}
-
-// attachRBACTokens stamps the per-(node, route) mTLS RBAC check token onto
-// every route that runs a forward_auth role check. The token proves to the
-// panel that the check subrequest comes from a node the control plane actually
-// placed the route on, instead of from any host that happens to sit inside the
-// mesh or trusted-proxy CIDRs (MTLS-01). No key configured = no token: the
-// panel then falls back to its allow-list-only behaviour.
-func (s *Service) attachRBACTokens(built []caddyapi.Route, ids []int64, nodeID int64) {
-	if len(s.MTLSRBACKey) == 0 || nodeID <= 0 || len(ids) != len(built) {
-		return
-	}
-	for i := range built {
-		if len(built[i].MTLSPathRules) == 0 {
-			continue
-		}
-		built[i].RBACNodeID = nodeID
-		built[i].RBACToken = security.MTLSRBACToken(s.MTLSRBACKey, nodeID, ids[i])
-	}
-}
-
-// panelBaseURL extracts scheme://host from a full URL (e.g. AskURL).
-func panelBaseURL(askURL string) string {
-	if askURL == "" {
-		return ""
-	}
-	u, err := url.Parse(askURL)
-	if err != nil || u.Host == "" {
-		return ""
-	}
-	return u.Scheme + "://" + u.Host
-}
-
-// attachRouteUpstreams fills caddyapi.Route.Upstreams for the built routes via
-// a single IN(...) query over route_upstreams, ordered positionally so
-// weighted_round_robin weights stay aligned with the emitted dial order.
-func (s *Service) attachRouteUpstreams(ctx context.Context, built []caddyapi.Route, ids []int64) {
-	if len(ids) == 0 {
-		return
-	}
-	idx := make(map[int64]int, len(ids))
-	ph := make([]string, len(ids))
-	args := make([]any, len(ids))
-	for i, id := range ids {
-		idx[id] = i
-		ph[i] = "?"
-		args[i] = id
-	}
-	rows, err := s.DB.QueryContext(ctx,
-		`SELECT route_id, host, port, weight, COALESCE(max_requests,0), COALESCE(enabled,1)
-		   FROM route_upstreams
-		  WHERE route_id IN (`+strings.Join(ph, ",")+`)
-		  ORDER BY route_id, sort_order ASC, id ASC`, args...)
-	if err != nil {
-		if s.Logger != nil {
-			s.Logger.Warn("route_upstreams load failed; routes stay single-dial", "err", err)
-		}
-		return
-	}
-	defer rows.Close()
-	for rows.Next() {
-		var rid int64
-		var host string
-		var port, weight, maxReq int
-		var enabled bool
-		if err := rows.Scan(&rid, &host, &port, &weight, &maxReq, &enabled); err != nil {
-			continue
-		}
-		// Soft-disabled upstreams are excluded from the emitted pool.
-		if !enabled {
-			continue
-		}
-		if i, ok := idx[rid]; ok {
-			built[i].Upstreams = append(built[i].Upstreams, caddyapi.Upstream{
-				Host:        host,
-				Port:        port,
-				Weight:      weight,
-				MaxRequests: maxReq,
-			})
-		}
-	}
-}
-
-func (s *Service) attachLocationRules(ctx context.Context, built []caddyapi.Route, ids []int64) {
-	if len(ids) == 0 {
-		return
-	}
-	idx := make(map[int64]int, len(ids))
-	ph := make([]string, len(ids))
-	args := make([]any, len(ids))
-	for i, id := range ids {
-		idx[id] = i
-		ph[i] = "?"
-		args[i] = id
-	}
-	rows, err := s.DB.QueryContext(ctx,
-		`SELECT route_id, path_glob, action, COALESCE(upstream_host,''), COALESCE(upstream_port,0),
-		        upstream_scheme, COALESCE(redirect_url,''), COALESCE(redirect_code,308), COALESCE(rewrite_uri,'')
-		   FROM route_location_rules
-		  WHERE route_id IN (`+strings.Join(ph, ",")+`)
-		  ORDER BY route_id, sort_order ASC, id ASC`, args...)
-	if err != nil {
-		if s.Logger != nil {
-			s.Logger.Warn("route_location_rules load failed; path rules skipped", "err", err)
-		}
-		return
-	}
-	defer rows.Close()
-	for rows.Next() {
-		var rid int64
-		var rule caddyapi.LocationRule
-		if err := rows.Scan(&rid, &rule.Path, &rule.Action, &rule.UpstreamHost, &rule.UpstreamPort,
-			&rule.UpstreamScheme, &rule.RedirectURL, &rule.RedirectCode, &rule.RewriteURI); err != nil {
-			continue
-		}
-		if i, ok := idx[rid]; ok {
-			built[i].LocationRules = append(built[i].LocationRules, rule)
-		}
-	}
-}
-
-// buildStreamsForNode reads the stream_routes table for one node and
-// returns caddyapi.StreamRoute values ready for the L4 builder. Joins on
-// services for the backend_ip (admin-only; stream routes never expose this
-// to the customer). Also loads stream_upstreams and advanced columns added
-// in migration 00061. Every destination is re-screened here, so a row stored
-// before a target became control-plane infrastructure cannot be re-emitted.
-func (s *Service) buildStreamsForNode(ctx context.Context, nodeID int64) ([]caddyapi.StreamRoute, error) {
-	rows, err := s.DB.QueryContext(ctx,
-		`SELECT sr.id, sr.protocol, sr.listen_port, sr.upstream_port,
-		        COALESCE(NULLIF(sr.backend_ip_override,''), sv.backend_ip),
-		        COALESCE(sr.match_mode,'any'),
-		        COALESCE(sr.match_values,''),
-		        COALESCE(sr.lb_policy,'round_robin'),
-		        COALESCE(sr.proxy_proto_in,'none'),
-		        COALESCE(sr.proxy_proto_out,'none'),
-		        COALESCE(sr.cidr_allow,''),
-		        COALESCE(sr.cidr_deny,'')
-		 FROM stream_routes sr JOIN services sv ON sv.id = sr.service_id
-		 WHERE sr.caddy_node_id = ? AND sr.status = 'active' AND sr.quarantined_at IS NULL
-		 ORDER BY sr.listen_port ASC`, nodeID)
-	if err != nil {
-		return nil, err
-	}
-	defer rows.Close()
-	var out []caddyapi.StreamRoute
-	for rows.Next() {
-		var r caddyapi.StreamRoute
-		var matchValuesCSV, cidrAllow, cidrDeny string
-		if err := rows.Scan(
-			&r.ID, &r.Protocol, &r.ListenPort, &r.UpstreamPort, &r.UpstreamIP,
-			&r.MatchMode, &matchValuesCSV, &r.LBPolicy,
-			&r.ProxyProtoIn, &r.ProxyProtoOut,
-			&cidrAllow, &cidrDeny,
-		); err != nil {
-			continue
-		}
-		r.MatchValues = splitCSV(matchValuesCSV)
-		r.CIDRAllow = splitCSV(cidrAllow)
-		r.CIDRDeny = splitCSV(cidrDeny)
-		out = append(out, r)
-	}
-	if len(out) == 0 {
-		return out, nil
-	}
-	// Attach multi-upstreams; routes without rows keep the legacy single-upstream.
-	s.attachStreamUpstreams(ctx, out)
-	return s.screenStreams(ctx, out)
-}
-
-// screenStreams re-validates every stream destination at emission time and
-// quarantines rows that now point at node or control-plane addresses. A deny
-// set that cannot be loaded fails the whole push closed.
-func (s *Service) screenStreams(ctx context.Context, in []caddyapi.StreamRoute) ([]caddyapi.StreamRoute, error) {
-	infra, err := streamguard.LoadInfraTargets(ctx, s.DB)
-	if err != nil {
-		return nil, fmt.Errorf("stream target screening unavailable: %w", err)
-	}
-	out, rejected := screenStreamSet(ctx, infra, in)
-	for _, rej := range rejected {
-		s.quarantineStream(ctx, rej.route, rej.cause)
-	}
-	return out, nil
-}
-
-// streamReject pairs a rejected stream with why it was rejected.
-type streamReject struct {
-	route caddyapi.StreamRoute
-	cause error
-}
-
-// screenStreamSet splits streams into emittable (destinations pinned to a
-// validated literal) and rejected. Pure, so the policy is testable without a DB.
-func screenStreamSet(ctx context.Context, infra *streamguard.InfraTargets, in []caddyapi.StreamRoute) ([]caddyapi.StreamRoute, []streamReject) {
-	out := make([]caddyapi.StreamRoute, 0, len(in))
-	var rejected []streamReject
-	for _, r := range in {
-		pinnedIP, perr := infra.ScreenAndPin(ctx, r.UpstreamIP, r.UpstreamPort)
-		if perr != nil {
-			rejected = append(rejected, streamReject{r, perr})
-			continue
-		}
-		r.UpstreamIP = pinnedIP
-		var bad error
-		for i, u := range r.Upstreams {
-			pinned, uerr := infra.ScreenAndPinAddress(ctx, u.Address)
-			if uerr != nil {
-				bad = uerr
-				break
-			}
-			r.Upstreams[i].Address = pinned
-		}
-		if bad != nil {
-			rejected = append(rejected, streamReject{r, bad})
-			continue
-		}
-		out = append(out, r)
-	}
-	return out, rejected
-}
-
-// quarantineStream parks an unsafe stream so it stops being re-emitted, and
-// leaves a visible trail instead of silently dropping it.
-func (s *Service) quarantineStream(ctx context.Context, r caddyapi.StreamRoute, cause error) {
-	reason := cause.Error()
-	if len(reason) > 255 {
-		reason = reason[:255]
-	}
-	if _, err := s.DB.ExecContext(ctx,
-		`UPDATE stream_routes SET quarantined_at = `+store.Now()+`, quarantine_reason = ? WHERE id = ?`,
-		reason, r.ID); err != nil && s.Logger != nil {
-		s.Logger.Error("stream quarantine flag failed", "stream_id", r.ID, "err", err)
-	}
-	if s.Logger != nil {
-		s.Logger.Warn("stream quarantined: unsafe destination", "stream_id", r.ID,
-			"listen_port", r.ListenPort, "reason", reason)
-	}
-	audit.Write(ctx, s.DB, s.Logger, nil, audit.Entry{
-		ActorType: audit.ActorSystem,
-		Action:    "stream.quarantined",
-		Entity:    "stream_route",
-		EntityID:  fmt.Sprintf("%d", r.ID),
-		Meta:      map[string]any{"reason": reason, "listen_port": r.ListenPort},
-	})
-}
-
-// attachStreamUpstreams loads stream_upstreams rows and maps them onto the
-// built StreamRoute slice by ID. Routes without upstream rows use the
-// legacy UpstreamIP:UpstreamPort path.
-func (s *Service) attachStreamUpstreams(ctx context.Context, built []caddyapi.StreamRoute) {
-	if len(built) == 0 {
-		return
-	}
-	// Collect IDs for IN(...) query - avoids N+1.
-	ids := make([]int64, len(built))
-	idx := make(map[int64]int, len(built))
-	for i, r := range built {
-		ids[i] = r.ID
-		idx[r.ID] = i
-	}
-
-	// Build parameterized IN clause.
-	placeholders := strings.Repeat("?,", len(ids))
-	placeholders = placeholders[:len(placeholders)-1]
-	args := make([]any, len(ids))
-	for i, id := range ids {
-		args[i] = id
-	}
-
-	rows, err := s.DB.QueryContext(ctx,
-		`SELECT stream_route_id, address, weight
-		 FROM stream_upstreams
-		 WHERE stream_route_id IN (`+placeholders+`)
-		 ORDER BY stream_route_id ASC, sort_order ASC, id ASC`, args...)
-	if err != nil {
-		s.Logger.Warn("stream_upstreams load failed; routes stay single-dial", "err", err)
-		return
-	}
-	defer rows.Close()
-	for rows.Next() {
-		var routeID int64
-		var u caddyapi.StreamUpstream
-		if err := rows.Scan(&routeID, &u.Address, &u.Weight); err != nil {
-			continue
-		}
-		if i, ok := idx[routeID]; ok {
-			built[i].Upstreams = append(built[i].Upstreams, u)
-		}
-	}
-}
-
-// splitCSV splits a comma-separated string into trimmed non-empty tokens.
-func splitCSV(s string) []string {
-	if s == "" {
-		return nil
-	}
-	parts := strings.Split(s, ",")
-	out := make([]string, 0, len(parts))
-	for _, p := range parts {
-		if t := strings.TrimSpace(p); t != "" {
-			out = append(out, t)
-		}
-	}
-	return out
-}
-
-// buildWildcardPolicies returns one WildcardPolicy per DISTINCT zone among
-// this node's active wildcard routes that has a dns_providers row. The
-// credential is decrypted here; a zone whose secret is missing/undecryptable
-// or whose provider is unsupported is SKIPPED (logged, zone only) so the node
-// never emits a DNS-01 policy without a working credential - which would fail
-// the entire /load. Returns nil when the gate is off (default).
-func (s *Service) buildWildcardPolicies(ctx context.Context, nodeID int64) []caddyapi.WildcardPolicy {
-	if !s.DNS01ModuleAvailable || s.DecryptSecret == nil {
-		return nil
-	}
-	rows, err := s.DB.QueryContext(ctx,
-		`SELECT DISTINCT dp.name, dp.provider, dp.api_token_enc
-		   FROM routes r
-		   JOIN dns_providers dp ON dp.name = r.wildcard_zone
-		  WHERE r.caddy_node_id = ?
-		    AND r.wildcard_enabled = 1
-		    AND r.status IN ('dns_ok','active','pending_ssl')
-		  ORDER BY dp.name ASC`, nodeID)
-	if err != nil {
-		return nil
-	}
-	defer rows.Close()
-	var out []caddyapi.WildcardPolicy
-	for rows.Next() {
-		var zone, provider, enc string
-		if err := rows.Scan(&zone, &provider, &enc); err != nil {
-			continue
-		}
-		if _, ok := caddyapi.DNSProviderBySlug(provider); !ok {
-			s.Logger.Warn("wildcard: unsupported provider, skipping", "zone", zone, "provider", provider)
-			continue
-		}
-		blob, derr := s.DecryptSecret(enc)
-		if derr != nil || blob == "" {
-			s.Logger.Error("wildcard: credential decrypt failed, skipping zone", "zone", zone)
-			continue
-		}
-		// Decode the JSON field map; legacy cloudflare rows hold a bare token.
-		fields := caddyapi.DecodeDNSFields(provider, blob)
-		if len(fields) == 0 {
-			s.Logger.Error("wildcard: credential blob unusable, skipping zone", "zone", zone)
-			continue
-		}
-		out = append(out, caddyapi.WildcardPolicy{Zone: zone, Provider: provider, Fields: fields})
-	}
-	return out
-}
-
-// buildManualCertsForNode returns operator-imported certs to load into this
-// node's Caddy cert pool: every manual_cert linked to a route this node serves
-// (direct caddy_node_id or a route_node_assignments fan-out), with its private
-// key decrypted. Same node-membership + status filter as buildRoutesForNode so
-// a cert is only shipped when its route is actually emitted. Unlinked certs
-// (route_id NULL) are stored-only and never pushed - the link is the binding.
-func (s *Service) buildManualCertsForNode(ctx context.Context, nodeID int64) []caddyapi.ManualCertPEM {
-	if s.DB == nil || s.DecryptSecret == nil {
-		return nil
-	}
-	rows, err := s.DB.QueryContext(ctx,
-		`SELECT mc.cert_pem, COALESCE(mc.chain_pem,''), mc.key_pem_enc
-		   FROM manual_certs mc
-		   JOIN routes r ON r.id = mc.route_id
-		  WHERE (r.caddy_node_id = ?
-		         OR EXISTS (SELECT 1 FROM route_node_assignments rna
-		                     WHERE rna.route_id = r.id AND rna.node_id = ?))
-		    AND r.ssl_enabled = 1
-		    AND r.status IN ('dns_ok','active','pending_ssl')
-		  ORDER BY mc.id ASC`, nodeID, nodeID)
-	if err != nil {
-		s.Logger.Error("manual certs: query failed", "node_id", nodeID, "err", err)
-		return nil
-	}
-	defer rows.Close()
-	var out []caddyapi.ManualCertPEM
-	for rows.Next() {
-		var certPEM, chainPEM, keyEnc string
-		if err := rows.Scan(&certPEM, &chainPEM, &keyEnc); err != nil {
-			continue
-		}
-		key, derr := s.DecryptSecret(keyEnc)
-		if derr != nil || key == "" {
-			// Never emit a cert with no key: Caddy would reject the whole /load.
-			s.Logger.Error("manual certs: key decrypt failed, skipping", "node_id", nodeID)
-			continue
-		}
-		// Caddy's load_pem wants leaf + chain in one PEM blob.
-		bundle := strings.TrimRight(certPEM, "\n")
-		if c := strings.TrimSpace(chainPEM); c != "" {
-			bundle += "\n" + c
-		}
-		out = append(out, caddyapi.ManualCertPEM{CertPEM: bundle, KeyPEM: key})
-	}
-	return out
-}
-
-// routeIsWildcard reports whether routeID has wildcard DNS-01 enabled. Used to
-// force a full /load (policy set lives outside the incremental @id surface).
-// Best-effort: any error returns false (caller proceeds with the normal path).
-func (s *Service) routeIsWildcard(ctx context.Context, routeID int64) bool {
-	var enabled bool
-	if err := s.DB.QueryRowContext(ctx,
-		"SELECT wildcard_enabled FROM routes WHERE id = ?", routeID).Scan(&enabled); err != nil {
-		return false
-	}
-	return enabled
-}
-
-// hashRoutes returns a stable SHA-256 over the deterministic JSON shape Caddy
-// would receive for these routes. Order is fixed by buildRoutesForNode.
-func hashRoutes(rs []caddyapi.Route) string {
-	objs := make([]map[string]any, 0, len(rs))
-	for _, r := range rs {
-		objs = append(objs, caddyapi.BuildRoute(r))
-	}
-	b, _ := json.Marshal(objs)
-	return hashBytes(b)
-}
-
-func hashBytes(b []byte) string {
-	sum := sha256.Sum256(b)
-	return hex.EncodeToString(sum[:])
-}
-
-// canonHashBytes unmarshals JSON then re-marshals so the hash is stable
-// across Caddy admin GET reformatting (Go map keys sort on Marshal).
-// Uses json.Decoder + UseNumber so port/ID values above 2^53 keep
-// integer precision; the default float64 path otherwise flaps the
-// hash and triggers infinite drift resync.
-// filterVirtualRoutes drops infra routes (panel self-route, wstunnel WSS route)
-// from a Caddy srv0/routes array so drift compares only customer routes, which
-// is all expectedNodeHash builds. BuildRoute emits @id="route_"+ID, so the panel
-// route (ID "panel_self") lands as "route_panel_self"; the wstunnel route is
-// built directly as "hpg_wstunnel_*". Customer routes are "route_<numeric>" and
-// are kept. Leaves input untouched if it's not the expected array.
-func filterVirtualRoutes(raw []byte) []byte {
-	var arr []json.RawMessage
-	if err := json.Unmarshal(raw, &arr); err != nil {
-		return raw
-	}
-	out := arr[:0]
-	for _, r := range arr {
-		var probe struct {
-			ID string `json:"@id"`
-		}
-		_ = json.Unmarshal(r, &probe)
-		if probe.ID == "route_panel_self" || strings.HasPrefix(probe.ID, "hpg_") {
-			continue
-		}
-		out = append(out, r)
-	}
-	b, err := json.Marshal(out)
-	if err != nil {
-		return raw
-	}
-	return b
-}
-
-func canonHashBytes(b []byte) string {
-	dec := json.NewDecoder(bytes.NewReader(b))
-	dec.UseNumber()
-	var v any
-	if err := dec.Decode(&v); err != nil {
-		return hashBytes(b)
-	}
-	canon, err := json.Marshal(v)
-	if err != nil {
-		return hashBytes(b)
-	}
-	return hashBytes(canon)
-}
-
-// expectedNodeHash computes the canonical-format hash of the Caddy routes
-// array Caddy would currently expose, derived from the DB. The drift probe
-// compares this to whatever Caddy actually returns over the admin API.
-func (s *Service) expectedNodeHash(ctx context.Context, nodeID int64) (string, error) {
-	built, _, err := s.buildRoutesForNode(ctx, nodeID)
-	if err != nil {
-		return "", err
-	}
-	return hashRoutes(built), nil
-}
-
-// ensureStableHash is a helper used in tests; not called from production.
-func ensureStableHash(rs []caddyapi.Route) string {
-	dup := make([]caddyapi.Route, len(rs))
-	copy(dup, rs)
-	sort.Slice(dup, func(i, j int) bool { return dup[i].ID < dup[j].ID })
-	return hashRoutes(dup)
-}
-
-// splitHostList explodes a stored comma/space separated hostname list.
-func splitHostList(raw string) []string {
-	out := []string{}
-	for _, p := range strings.FieldsFunc(raw, func(r rune) bool {
-		return r == ',' || r == ' ' || r == '\t' || r == '\n' || r == '\r' || r == ';'
-	}) {
-		if v := strings.ToLower(strings.TrimSpace(p)); v != "" {
-			out = append(out, v)
-		}
-	}
-	return out
-}
-
-// newVerifyToken returns a 32-hex-char (128-bit) random nonce the domain owner
-// publishes as a TXT record to prove control. Hex so it is DNS-TXT-safe.
-func newVerifyToken() (string, error) {
-	b := make([]byte, 16)
-	if _, err := rand.Read(b); err != nil {
-		return "", err
-	}
-	return hex.EncodeToString(b), nil
-}
-
-func validDomain(d string) bool {
-	if len(d) == 0 || len(d) > 253 {
-		return false
-	}
-	if net.ParseIP(d) != nil {
-		return false // IPs not allowed as hostnames
-	}
-	if strings.Contains(d, "..") {
-		return false
-	}
-	if !strings.Contains(d, ".") {
-		return false
-	}
-	for _, c := range d {
-		switch {
-		case c >= 'a' && c <= 'z',
-			c >= 'A' && c <= 'Z',
-			c >= '0' && c <= '9',
-			c == '-', c == '.':
-		default:
-			return false
-		}
-	}
-	// Each DNS label: 1-63 chars, no leading or trailing hyphen (RFC 1035 §2.3.1).
-	for _, label := range strings.Split(d, ".") {
-		if len(label) == 0 || len(label) > 63 {
-			return false
-		}
-		if label[0] == '-' || label[len(label)-1] == '-' {
-			return false
-		}
-	}
-	return true
-}
-
-// ValidDomain exposes the internal hostname validator so handlers editing a
-// route (not via Create) enforce the same shape check.
-func ValidDomain(d string) bool { return validDomain(d) }
-
-// NewVerifyToken exposes the DNS-TXT ownership nonce generator so an
-// out-of-service domain change can re-arm verification.
-func NewVerifyToken() (string, error) { return newVerifyToken() }
-
-// looksLikeIP reports whether s parses as IPv4/IPv6 (not a hostname).
-func looksLikeIP(s string) bool {
-	return net.ParseIP(strings.TrimSpace(s)) != nil
-}
-
-func truncErr(e error) string {
-	s := e.Error()
-	if len(s) > 240 {
-		s = s[:240] + "..."
-	}
-	return s
 }
